% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractModule-class.R
\name{AbstractModule}
\alias{AbstractModule}
\title{Abstract module}
\description{
An abstract module.
}
\section{Super class}{
\code{\link[nitro:QueueGenerator]{nitro::QueueGenerator}} -> \code{AbstractModule}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{The name of the module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractModule-clone}{\code{AbstractModule$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nitro" data-topic="QueueGenerator" data-id="queue"><a href='../../nitro/html/QueueGenerator.html#method-QueueGenerator-queue'><code>nitro::QueueGenerator$queue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
