% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractResamplingOptions-class.R
\name{AbstractResamplingOptions}
\alias{AbstractResamplingOptions}
\title{Define options common to resampling analyses}
\description{
\code{AbstractResamplingOptions} is an R6 class that defines options common
  to resampling analyses.
}
\section{Super class}{
\code{\link[nitro:AbstractExtension]{nitro::AbstractExtension}} -> \code{AbstractResamplingOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{replications}}{An integer value indicating the number of resampling
replications to perform.}

\item{\code{cutoff}}{An integer value indicating the cutoff value for
frequencies.}

\item{\code{frequency_summary}}{A character vector indicating which method(s) to
  use to summarize supports. More than one option can be specified. The
  options are:
\itemize{
  \item \code{difference}: frequency differences (i.e., group supported/contradicted), default;
  \item \code{absolute}: absolute frequencies; or
  \item \code{slope}: frequency slopes.
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractResamplingOptions-new}{\code{AbstractResamplingOptions$new()}}
\item \href{#method-AbstractResamplingOptions-clone}{\code{AbstractResamplingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractResamplingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractResamplingOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResamplingOptions$new(
  replications = 100,
  frequency_summary = "absolute"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replications}}{An integer value indicating the number of resampling
replications to perform.}

\item{\code{frequency_summary}}{A character vector indicating which method(s) to
  use to summarize supports. More than one option can be specified. The
  options are:
\itemize{
  \item \code{difference}: frequency differences (i.e., group supported/contradicted), detault;
  \item \code{absolute}: absolute frequencies; or
  \item \code{slope}: frequency slopes.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractResamplingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractResamplingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResamplingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
