% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractSupportOptions-class.R
\name{AbstractSupportOptions}
\alias{AbstractSupportOptions}
\title{Define options common to resampling analyses}
\description{
\code{AbstractSupportOptions} is an R6 class that defines options common
  to resampling analyses.
}
\section{Super classes}{
\code{\link[nitro:QueueGenerator]{nitro::QueueGenerator}} -> \code{\link[nitro:AbstractModule]{nitro::AbstractModule}} -> \code{AbstractSupportOptions}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractSupportOptions-queue}{\code{AbstractSupportOptions$queue()}}
\item \href{#method-AbstractSupportOptions-clone}{\code{AbstractSupportOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractSupportOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractSupportOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractSupportOptions$queue(.queue = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.queue}}{A \code{\link{QueueGenerator}} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractSupportOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractSupportOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractSupportOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
