% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractTreeSearchOptions-class.R
\name{AbstractTreeSearchOptions}
\alias{AbstractTreeSearchOptions}
\title{Abstract tree search options}
\description{
\code{AbstractTreeSearchOptions} is an R6 class that defines properties
  common to tree searches. Intended to be inherited, not instantiated
  directly.
}
\section{Super classes}{
\code{\link[nitro:QueueGenerator]{nitro::QueueGenerator}} -> \code{\link[nitro:AbstractModule]{nitro::AbstractModule}} -> \code{AbstractTreeSearchOptions}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractTreeSearchOptions-clone}{\code{AbstractTreeSearchOptions$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nitro" data-topic="QueueGenerator" data-id="queue"><a href='../../nitro/html/QueueGenerator.html#method-QueueGenerator-queue'><code>nitro::QueueGenerator$queue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractTreeSearchOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractTreeSearchOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractTreeSearchOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
