% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintsBase-methods.R
\name{addConstraint}
\alias{addConstraint}
\title{Add new constraint}
\usage{
addConstraint(
  tree_search,
  fixed_taxa,
  floating_taxa = character(),
  type = c("positive", "negative")
)
}
\arguments{
\item{tree_search}{an object inheriting
\code{"\linkS4class{NitroTreeSearch}"}}

\item{fixed_taxa}{a character vector of taxa to set as fixed constraints}

\item{floating_taxa}{a character vector of taxa to set as floating
constraints}

\item{type}{a character vector indicating whether the constraint will be
'\code{positive}' or '\code{negative}'}
}
\description{
Function to add a new constraint to a constraint group.
}
