% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchBreakingOptions-class.R
\name{BranchBreakingOptions}
\alias{BranchBreakingOptions}
\title{Set options for branch breaking}
\description{
\code{BranchBreakingOptions} is an R6 class that defines the set of optiopns
required to perform branch breaking.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{swapper}}{A character vector indicating the branch swapping algorithm
to use. Options are either '\code{tbr}' (the default) or '\code{spr}'.}

\item{\code{cluster_size}}{An integer value indicating the number of nodes
(clusters) to use during swapping. Applied only when \code{swapper} is
set to \code{tbr}. Larger clusters will result in faster swapping as
matrix size increases. If left as 0 (the default), cluster size will be
automatically determined.}

\item{\code{safe_unclip}}{A logical value indicating whether to use a safer but
slower method for updating buffers when clipping to find a better tree.
Applied only when \code{swapper} is set to \code{tbr}.}

\item{\code{fill_only}}{A logical value indicating whether to stop swapping when
the tree buffer is full.}

\item{\code{save_multiple}}{A logical value indicating whether to save multiple
trees during swapping.}

\item{\code{random_clip}}{A logical value indicating whether to randomize the
tree clipping sequence.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BranchBreakingOptions-new}{\code{BranchBreakingOptions$new()}}
\item \href{#method-BranchBreakingOptions-print}{\code{BranchBreakingOptions$print()}}
\item \href{#method-BranchBreakingOptions-queue}{\code{BranchBreakingOptions$queue()}}
\item \href{#method-BranchBreakingOptions-clone}{\code{BranchBreakingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchBreakingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-BranchBreakingOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchBreakingOptions$new(
  swapper = "tbr",
  cluster_size = 0,
  safe_unclip = FALSE,
  fill_only = FALSE,
  save_multiple = TRUE,
  random_clip = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{swapper}}{A character vector indicating the branch swapping algorithm
to use. Options are either '\code{tbr}' (the default) or '\code{spr}'.}

\item{\code{cluster_size}}{An integer value indicating the number of nodes
(clusters) to use during swapping. Applied only when \code{swapper} is
set to \code{tbr}. Larger clusters will result in faster swapping as
matrix size increases. If left as 0 (the default), cluster size will be
automatically determined.}

\item{\code{safe_unclip}}{A logical value indicating whether to use a safer but
slower method for updating buffers when clipping to find a better tree.
Applied only when \code{swapper} is set to \code{tbr}.}

\item{\code{fill_only}}{A logical value indicating whether to stop swapping when
the tree buffer is full.}

\item{\code{save_multiple}}{A logical value indicating whether to save multiple
trees during swapping.}

\item{\code{random_clip}}{A logical value indicating whether to randomize the
tree clipping sequence.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchBreakingOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-BranchBreakingOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchBreakingOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchBreakingOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-BranchBreakingOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchBreakingOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchBreakingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BranchBreakingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchBreakingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
