% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchSupportOptions-class.R
\name{BranchSupportOptions}
\alias{BranchSupportOptions}
\title{Set options for branch support calculation}
\description{
\code{BranchSupportOptions} is an R6 class that defines options for calculating
  decay indices.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{index_type}}{A character vector indicating the type of index to
calculate. Options are either \code{absolute} (default) or
\code{relative}.}

\item{\code{group_loss_cost}}{A character vector indicating what cost to use
for losing a group in a suboptimal tree.}

\item{\code{suboptimal_steps}}{An integer value indicating the maximum absolute
fit difference (i.e., length difference) of suboptimal trees to use
for calculating supports.}

\item{\code{relative_suboptimal_fit}}{A numeric value indicating the maximum
relative fit difference of suboptimal trees to use for calculating
supports.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BranchSupportOptions-new}{\code{BranchSupportOptions$new()}}
\item \href{#method-BranchSupportOptions-print}{\code{BranchSupportOptions$print()}}
\item \href{#method-BranchSupportOptions-queue}{\code{BranchSupportOptions$queue()}}
\item \href{#method-BranchSupportOptions-clone}{\code{BranchSupportOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSupportOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSupportOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSupportOptions$new(
  index_type = "absolute",
  group_loss_cost = "standard",
  suboptimal_steps = 10,
  relative_suboptimal_fit = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index_type}}{A character vector indicating the type of index to
calculate. Options are either \code{absolute} (default) or
\code{relative}.}

\item{\code{group_loss_cost}}{A character vector indicating what cost to use
for losing a group in a suboptimal tree.}

\item{\code{suboptimal_steps}}{An integer value indicating the maximum length
of suboptimal trees to keep during branch swapping.}

\item{\code{relative_suboptimal_fit}}{A numeric value indicating the maximum
relative fit difference of suboptimal trees to use for calculating
supports.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSupportOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSupportOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSupportOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSupportOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSupportOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSupportOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSupportOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSupportOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSupportOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
