% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchswap.R
\name{branchswap}
\alias{branchswap}
\title{Phylogenetic analysis using branch swapping}
\usage{
branchswap(tnt.path, matrix, hold = 100, outgroup = NULL,
  replications = 10, hold.rep = 10, keepall = FALSE)
}
\arguments{
\item{tnt.path}{The location of the TNT command-line binary.}

\item{matrix}{A \code{phyDat} object of the matrix.}

\item{hold}{The maximum number of trees to allow TNT to hold in memory (typically \code{replications} * \code{hold.rep}).}

\item{outgroup}{The outgroup taxon for the phylogenetic analysis. By default, the first taxon in the matrix is considered the outgroup.}

\item{replications}{The number of branch swapping replications.}

\item{hold.rep}{The maximum number of trees to retain during each replication.}

\item{keepall}{Retain all generated trees from each replication regardless of length.}
}
\value{
A \code{multiPhylo} object.
}
\description{
Phylogenetic analysis using branch swapping
}
