% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchSwap-class.R
\name{BranchSwappingOptions}
\alias{BranchSwappingOptions}
\title{Define a branch swapping analysis}
\description{
\code{BranchSwappingOptions} is an R6 class that defines the set of options
for performing a branch sweapping phylogenetic analysis with replications
in \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{replications}}{An integer value indicating the number of replications.}

\item{\code{hold_rep}}{An integer value indicating the maximum number of trees to
retain during each replication.}

\item{\code{keep_all}}{A logical value indicating whether to retain all generated
trees from each replication regardless of length.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BranchSwappingOptions-new}{\code{BranchSwappingOptions$new()}}
\item \href{#method-BranchSwappingOptions-print}{\code{BranchSwappingOptions$print()}}
\item \href{#method-BranchSwappingOptions-queue}{\code{BranchSwappingOptions$queue()}}
\item \href{#method-BranchSwappingOptions-clone}{\code{BranchSwappingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSwappingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSwappingOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSwappingOptions$new(replications = 10, hold_rep = 10, keep_all = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replications}}{An integer value indicating the number of replications.}

\item{\code{hold_rep}}{An integer value indicating the maximum number of trees to
retain during each replication.}

\item{\code{keep_all}}{A logical value indicating whether to retain all generated
trees from each replication regardless of length.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSwappingOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSwappingOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSwappingOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSwappingOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSwappingOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSwappingOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BranchSwappingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BranchSwappingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BranchSwappingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
