% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCharacterMatrix.R
\name{c.AbstractCharacterMatrix}
\alias{c.AbstractCharacterMatrix}
\title{Concatenate character matrices}
\usage{
\method{c}{AbstractCharacterMatrix}(...)
}
\arguments{
\item{...}{Character matrix objects.}
}
\description{
Concatenate character matrix objects into a list.
}
