% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicCommand-class.R
\name{c.CommandList}
\alias{c.CommandList}
\title{Combine CommandList Objects}
\usage{
\method{c}{CommandList}(...)
}
\arguments{
\item{...}{One or more \code{CommandList} objects.}
}
\value{
A \code{CommandList} object.
}
\description{
Flattens and combines one or more \code{CommandList} objects into a single
\code{CommandList}.
}
