% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{ordered_characters}
\alias{ordered_characters}
\alias{ordered_characters,NitroTreeSearch-method}
\alias{ordered_characters<-}
\alias{ordered_characters<-,NitroTreeSearch,numeric-method}
\alias{ordered_characters<-,NitroTreeSearch,NULL-method}
\alias{inactive_characters}
\alias{inactive_characters,NitroTreeSearch-method}
\alias{inactive_characters<-}
\alias{inactive_characters<-,NitroTreeSearch,numeric-method}
\alias{inactive_characters<-,NitroTreeSearch,NULL-method}
\title{Character modification functions}
\usage{
ordered_characters(n)

\S4method{ordered_characters}{NitroTreeSearch}(n)

ordered_characters(n) <- value

\S4method{ordered_characters}{NitroTreeSearch,numeric}(n) <- value

\S4method{ordered_characters}{NitroTreeSearch,`NULL`}(n) <- value

inactive_characters(n)

\S4method{inactive_characters}{NitroTreeSearch}(n)

inactive_characters(n) <- value

\S4method{inactive_characters}{NitroTreeSearch,numeric}(n) <- value

\S4method{inactive_characters}{NitroTreeSearch,`NULL`}(n) <- value
}
\arguments{
\item{n}{a \code{"\linkS4class{NitroTreeSearch}"} object.}

\item{value}{a numeric vector indicating the characters to be marked as inactive.}
}
\value{
a numeric vector indicating the characters defined as ordered.

a numeric vector indicating the characters defined as inactive.
}
\description{
List ordered characters

List inactive characters
}
\details{
Function to return or set indexes of characters as ordered.

Function that returns the indexes of characters defined as inactive.
}
