% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-classes.R
\name{check_NitroBranchSwap}
\alias{check_NitroBranchSwap}
\title{Validate Nitro S4 objects}
\usage{
check_NitroBranchSwap(object)
}
\arguments{
\item{object}{a prospective object that inherits \code{NitroBranchSwap}.}
}
\value{
a validated object.
}
\description{
Validate Nitro S4 objects
}
