% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-classes.R
\name{check_NitroDriven}
\alias{check_NitroDriven}
\title{Validate Nitro S4 objects}
\usage{
check_NitroDriven(object)
}
\arguments{
\item{object}{a prospective object that inherits \code{NitroDriven}.}
}
\value{
a validated object.
}
\description{
Validate Nitro S4 objects
}
