% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-classes.R
\name{check_NitroTreeSearch}
\alias{check_NitroTreeSearch}
\title{Validate Nitro S4 objects}
\usage{
check_NitroTreeSearch(object)
}
\arguments{
\item{object}{a prospective object that inherits
\code{"\linkS4class{NitroTreeSearch}"}.}
}
\value{
a validated object.
}
\description{
Validate Nitro S4 objects
}
