% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{collapse}
\alias{collapse}
\alias{collapse,NitroTreeSearch-method}
\alias{collapse<-}
\alias{collapse<-,NitroTreeSearch,numeric-method}
\title{Zero-length branch collapse rule}
\usage{
collapse(n)

\S4method{collapse}{NitroTreeSearch}(n)

collapse(n) <- value

\S4method{collapse}{NitroTreeSearch,numeric}(n) <- value
}
\arguments{
\item{n}{a \code{"\linkS4class{NitroTreeSearch}"} object.}

\item{value}{an integer indicating the rule for collapsing of zero length branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}
}
\value{
a numeric vector indicating the rule used for collapsing of
zero-length branches.
}
\description{
Function to return or set the rule used for collapsing of zero-length
branches.
}
