% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstraintedSectorialSearchOptions-class.R
\name{ConstrainedSectorialSearchOptions}
\alias{ConstrainedSectorialSearchOptions}
\title{Set options for a constrained sectorial search}
\description{
\code{ConstrainedSectorialSearchOptions} is an R6 class that defines the set
  of options required to perform sectorial searching analyses in
  \code{nitro}.
}
\section{Super class}{
\code{\link[nitro:SectorialSearchBaseOptions]{nitro::SectorialSearchBaseOptions}} -> \code{ConstrainedSectorialSearchOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{min_fork}}{An integer value indicating the minimum fork number to use
with constraint-based sectorial searches.}

\item{\code{max_fork}}{An integer value indicating the maximum fork number to use
with constraint-based sectorial searches.}

\item{\code{rounds}}{An integer value indicating the number of times to cycle over
groups in constraint-based selections.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConstrainedSectorialSearchOptions-new}{\code{ConstrainedSectorialSearchOptions$new()}}
\item \href{#method-ConstrainedSectorialSearchOptions-print}{\code{ConstrainedSectorialSearchOptions$print()}}
\item \href{#method-ConstrainedSectorialSearchOptions-queue}{\code{ConstrainedSectorialSearchOptions$queue()}}
\item \href{#method-ConstrainedSectorialSearchOptions-clone}{\code{ConstrainedSectorialSearchOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConstrainedSectorialSearchOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConstrainedSectorialSearchOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstrainedSectorialSearchOptions$new(
  min_fork = 10,
  max_fork = 10,
  rounds = 3,
  buffer = TRUE,
  slack = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_fork}}{An integer value indicating the minimum fork number to use
with constraint-based sectorial searches.}

\item{\code{max_fork}}{An integer value indicating the maximum fork number to use
with constraint-based sectorial searches.}

\item{\code{rounds}}{An integer value indicating the number of times to cycle over
groups in constraint-based selections.}

\item{\code{buffer}}{A logical value indicating whether to use an independent
memory buffer for analysis of sectors.}

\item{\code{slack}}{An integer value indicating the percentage to increase the
available memory during searches.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConstrainedSectorialSearchOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-ConstrainedSectorialSearchOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstrainedSectorialSearchOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConstrainedSectorialSearchOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-ConstrainedSectorialSearchOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstrainedSectorialSearchOptions$queue(set_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_only}}{A logical indicating whether to instruct the command to
execute immediately (\code{TRUE}) or set the variables for future
execution (\code{FALSE}, default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConstrainedSectorialSearchOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConstrainedSectorialSearchOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstrainedSectorialSearchOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
