% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstraintBaseOptions-class.R
\name{ConstraintBaseOptions}
\alias{ConstraintBaseOptions}
\title{Set constraints on monophyly}
\description{
\code{ConstraintBaseOptions} is an R6 class that sets common options for tree
  searches with constraints.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{is_positive}}{A logical value indicating whether the constraint is
positive (\code{TRUE}) or negative (\code{FALSE}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConstraintBaseOptions-new}{\code{ConstraintBaseOptions$new()}}
\item \href{#method-ConstraintBaseOptions-clone}{\code{ConstraintBaseOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConstraintBaseOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConstraintBaseOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstraintBaseOptions$new(is_positive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{is_positive}}{A logical value indicating whether the constraint is
positive (\code{TRUE}) or negative (\code{FALSE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConstraintBaseOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConstraintBaseOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstraintBaseOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
