% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Create a new character matrix}
\usage{
create_matrix(data, ...)
}
\arguments{
\item{data}{A \code{phyDat} or a \code{data.frame} object, to create a
discrete or continuous character-taxon matrix, respectively.}

\item{...}{Arguments to be passed on to the R6 class responsible for creating
the matrix object. See \code{\link{DiscreteMatrix}} and
\code{\link{ContinuousMatrix}} for details.}
}
\description{
Create a new discrete or continuous character-taxon matrix.
}
\seealso{
The \code{new} method of \code{\link{DiscreteMatrix}} and
  \code{\link{ContinuousMatrix}} describe optional arguments for matrix
  creation.
}
