% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_object.R
\name{create_new_object}
\alias{create_new_object}
\title{Create a new object}
\usage{
create_new_object(name, object_choice, ...)
}
\arguments{
\item{name}{The name of the object to create.}

\item{object_choice}{The possible options for object creation.}
}
\description{
Create a new R6 object from a set of possible options.
}
