% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driven.R
\name{driven}
\alias{driven}
\title{Phylogenetic analysis using branch swapping}
\usage{
driven(tnt.path, matrix, hold = 100, outgroup = NULL, hits = 1,
  replications = 4, autoconstrain = "wagner", rss = TRUE,
  css = FALSE, xss = FALSE, ratchet.cycles = NULL,
  drifting.cycles = 30, hybridization = FALSE, fusing.rounds = NULL,
  consense.times = NULL, multiply = FALSE, hold.rep = 1,
  keepall = FALSE)
}
\arguments{
\item{tnt.path}{The location of the TNT command-line binary.}

\item{matrix}{A \code{phyDat} object of the matrix.}

\item{hold}{The maximum number of trees to allow TNT to hold in memory.}

\item{outgroup}{The outgroup taxon for the phylogenetic analysis. By default, the first taxon in the matrix is considered the outgroup.}

\item{hits}{The number of times to reach the shortest tree for each replication.}

\item{replications}{The number of replications.}

\item{rss, css, xss}{Use random, constraint or exclusive sectorial searches.}

\item{ratchet.cycles}{The number of cycles of the parsimony ratchet.}

\item{drifting.cycles}{The number of cycles of tree drifting.}

\item{fusing.rounds}{The number of rounds of fusing.}

\item{consense.times}{Number of times to consense until the consensus is stablilised.}

\item{multiply}{After hitting target score, find additional trees by fusing suboptimal with optimal trees}

\item{hold.rep}{The maximum number of trees to retain during each replication.}

\item{keepall}{Retain all generated trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When \code{hits} = 1, it is trees from each of the
RAS + TBR +  SS or DFT or RAT, in addition to the trees resulting from fusing those.  When \code{hits} > 1,
then it means the trees resulting from fusing the initial starting trees for each of starting points.}
}
\value{
A \code{multiPhylo} object.
}
\description{
Phylogenetic analysis using branch swapping
}
