% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrivenSearchOptions-class.R
\name{DrivenSearchOptions}
\alias{DrivenSearchOptions}
\title{Set options for driven searches}
\description{
\code{DrivenSearchOptions} is an R6 class that defines the set of parameters
required to perform a driven search in \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{replications}}{An integer value indicating the number of replications.}

\item{\code{hits}}{An integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{\code{consense_times}}{An integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{\code{keep_all}}{A logical value indicating whether to retain all generated
trees from each replication regardless of length.}

\item{\code{multiply}}{A logical value indicating whether to find additional trees
by fusing suboptimal trees with optimal trees.}

\item{\code{sectorial_search}}{A list of objects of inheriting
\code{"\link{SectorialSearchOptions}"}.}

\item{\code{tree_fusing}}{An object of class \code{"\link{TreeFusingOptions}"}.}

\item{\code{tree_hybridizing}}{An object of class \code{"\link{TreeHybridizingOptions}"}.}

\item{\code{tree_drifting}}{An object of class \code{"\link{TreeDriftingOptions}"}.}

\item{\code{ratchet}}{An object of class \code{"\link{RatchetOptions}"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DrivenSearchOptions-new}{\code{DrivenSearchOptions$new()}}
\item \href{#method-DrivenSearchOptions-print}{\code{DrivenSearchOptions$print()}}
\item \href{#method-DrivenSearchOptions-queue}{\code{DrivenSearchOptions$queue()}}
\item \href{#method-DrivenSearchOptions-clone}{\code{DrivenSearchOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrivenSearchOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-DrivenSearchOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrivenSearchOptions$new(
  replications = 4,
  hits = 1,
  consense_times = 0,
  keep_all = FALSE,
  multiply = TRUE,
  sectorial_search = NULL,
  tree_fusing = NULL,
  tree_hybridizing = NULL,
  tree_drifting = NULL,
  ratchet = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replications}}{an integer value indicating the number of replications.}

\item{\code{hits}}{An integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{\code{consense_times}}{An integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{\code{keep_all}}{A logical value indicating whether to retain all generated
trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When
\code{hits} = 1, it is trees from each of the RAS + TBR +  SS or DFT or
RAT, in addition to the trees resulting from fusing those. When
\code{hits} > 1, then it means the trees resulting from fusing the
initial starting trees for each of starting points.}

\item{\code{multiply}}{A logical value indicating whether to find additional trees
by fusing suboptimal trees with optimal trees.}

\item{\code{sectorial_search}}{A list of objects of inheriting
\code{"\link{NitroSectorialSearch}"}.}

\item{\code{tree_fusing}}{An object of class \code{"\link{NitroTreeFusing}"}.}

\item{\code{tree_hybridizing}}{An object of class \code{"\link{NitroTreeHybridize}"}.}

\item{\code{tree_drifting}}{An object of class \code{"\link{NitroTreeDrifting}"}.}

\item{\code{ratchet}}{An object of class \code{"\link{NitroRatchet}"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrivenSearchOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-DrivenSearchOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrivenSearchOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrivenSearchOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-DrivenSearchOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrivenSearchOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrivenSearchOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrivenSearchOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrivenSearchOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
