% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoCommand-class.R
\name{EchoCommand}
\alias{EchoCommand}
\title{Echo Command}
\description{
An \link[R6:R6Class]{R6} class that configures whether TNT echoes commands
to the console during execution.

This command is created automatically as part of the standard
\link{CommandQueue} setup. Users do not typically need to instantiate it
directly.
}
\details{
\subsection{Command output}{

\verb{$render()} produces \verb{echo =;} (enabled) or \verb{echo -;} (disabled).
}

\subsection{Queue integration}{

Calling \verb{$enqueue()} adds this command to a \link{CommandQueue} at priority
\code{100}, placing it near the start of the execution sequence.
}
}
\keyword{internal}
\section{Super class}{
\code{\link[nitro:BasicCommand]{BasicCommand}} -> \code{EchoCommand}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{enable}}{[\code{logical(1)}]\cr
Whether to echo commands to the console during execution.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-EchoCommand-enqueue}{\code{EchoCommand$enqueue()}}
    \item \href{#method-EchoCommand-initialize}{\code{EchoCommand$new()}}
    \item \href{#method-EchoCommand-render}{\code{EchoCommand$render()}}
    \item \href{#method-EchoCommand-clone}{\code{EchoCommand$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="format"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-format'><code>BasicCommand$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="get_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-get_dependency'><code>BasicCommand$get_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="new_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-new_dependency'><code>BasicCommand$new_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="print"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-print'><code>BasicCommand$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="set_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-set_dependency'><code>BasicCommand$set_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="transform"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-transform'><code>BasicCommand$transform()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EchoCommand-enqueue"></a>}}
\if{latex}{\out{\hypertarget{method-EchoCommand-enqueue}{}}}
\subsection{\code{EchoCommand$enqueue()}}{
  Add this command to a \link{CommandQueue}.

Adds this command at priority \code{100} by default.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EchoCommand$enqueue(.queue, priority = 100)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{.queue}}{A \link{CommandQueue} object.}
      \item{\code{priority}}{[\code{integer(1)}]\cr
A non-negative integer controlling execution order. Lower values
execute first.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EchoCommand-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-EchoCommand-initialize}{}}}
\subsection{\code{EchoCommand$new()}}{
  Create a new \code{EchoCommand} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EchoCommand$new(enable)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{enable}}{[\code{logical(1)}]\cr
Whether to echo commands to the console. See the \verb{$enable} field.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{EchoCommand} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EchoCommand-render"></a>}}
\if{latex}{\out{\hypertarget{method-EchoCommand-render}{}}}
\subsection{\code{EchoCommand$render()}}{
  Render the TNT \code{echo} command string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EchoCommand$render(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Not used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A single-element character vector containing the TNT
command.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EchoCommand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EchoCommand-clone}{}}}
\subsection{\code{EchoCommand$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EchoCommand$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
