% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExclusiveSectorialSearchOptions-class.R
\name{ExclusiveSectorialSearchOptions}
\alias{ExclusiveSectorialSearchOptions}
\title{Set options for an exclusive sectorial search}
\description{
\code{ExclusiveSectorialSearchOptions} is an R6 class that defines the set
  of options required to perform exclusive sectorial searching analyses in
  \code{nitro}.
}
\section{Super class}{
\code{\link[nitro:SectorialSearchBaseOptions]{nitro::SectorialSearchBaseOptions}} -> \code{ExclusiveSectorialSearchOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{selections}}{An integer value indicating the number of exclusive (i.e.,
non-overlapping) sectors to subdivide the entire tree into.}

\item{\code{rounds}}{An integer value indicating the number of times to repeat
exclusive sector selection and analysis.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExclusiveSectorialSearchOptions-new}{\code{ExclusiveSectorialSearchOptions$new()}}
\item \href{#method-ExclusiveSectorialSearchOptions-print}{\code{ExclusiveSectorialSearchOptions$print()}}
\item \href{#method-ExclusiveSectorialSearchOptions-queue}{\code{ExclusiveSectorialSearchOptions$queue()}}
\item \href{#method-ExclusiveSectorialSearchOptions-clone}{\code{ExclusiveSectorialSearchOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExclusiveSectorialSearchOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExclusiveSectorialSearchOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExclusiveSectorialSearchOptions$new(
  selections,
  rounds = 2,
  buffer = TRUE,
  slack = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selections}}{An integer value indicating the number of exclusive (i.e.,
non-overlapping) sectors to subdivide the entire tree into.}

\item{\code{rounds}}{An integer value indicating the number of times to repeat
exclusive sector selection and analysis.}

\item{\code{buffer}}{A logical value indicating whether to use an independent
memory buffer for analysis of sectors.}

\item{\code{slack}}{An integer value indicating the percentage to increase the
available memory during searches.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExclusiveSectorialSearchOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExclusiveSectorialSearchOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExclusiveSectorialSearchOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExclusiveSectorialSearchOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-ExclusiveSectorialSearchOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExclusiveSectorialSearchOptions$queue(set_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_only}}{A logical indicating whether to instruct the command to
execute immediately (\code{TRUE}) or set the variables for future
execution (\code{FALSE}, default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExclusiveSectorialSearchOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExclusiveSectorialSearchOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExclusiveSectorialSearchOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
