% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroDriven-methods.R
\name{hits}
\alias{hits}
\alias{hits,NitroDriven-method}
\alias{hits<-}
\alias{hits<-,NitroDriven,numeric-method}
\title{Hits}
\usage{
hits(n)

\S4method{hits}{NitroDriven}(n)

hits(n) <- value

\S4method{hits}{NitroDriven,numeric}(n) <- value
}
\arguments{
\item{n}{an object of class \code{NitroDriven}.}

\item{value}{an integer indicating the number of times the shortest tree
must be found on consecutive re-runs of the analysis before stopping.}
}
\value{
an integer value indicating the number of hits.
}
\description{
Function that returns or sets the number of times that shortest tree(s) of
the same length must found on consecutive re-runs of the analysis before
stopping.
}
