% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{hold}
\alias{hold}
\alias{hold,NitroTreeSearch-method}
\alias{hold<-}
\alias{hold<-,NitroTreeSearch-method}
\title{Hold trees}
\usage{
hold(n)

\S4method{hold}{NitroTreeSearch}(n)

hold(n) <- value

\S4method{hold}{NitroTreeSearch}(n) <- value
}
\arguments{
\item{n}{a \code{"\linkS4class{NitroTreeSearch}"} object}

\item{value}{a numeric or integer indicating the number of trees to hold in
TNT's tree buffer.}
}
\value{
a numeric or integer indicating the number of trees to hold in TNT's
tree buffer.
}
\description{
A function that returns or sets the maximum number of trees to hold in TNT's
tree buffer.
}
