% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ienum.R
\name{implicit.enum}
\alias{implicit.enum}
\title{Phylogenetic analysis using implicit enumeration}
\usage{
implicit.enum(matrix, hold = 100, collapse = 3, outgroup = NULL)
}
\arguments{
\item{matrix}{a \code{phyDat} object of the matrix.}

\item{hold}{an integer value indicating the maximum number of trees to allow
TNT to hold in memory.}

\item{collapse}{an integer indicating the rule for collapsing of zero length
branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}

\item{outgroup}{the name of the taxon to set as the outgroup for the
phylogenetic analysis. By default, the first taxon in the matrix is the
outgroup.}
}
\value{
a list containing the search parameters and TNT command string, a
  \code{phyDat} object containing the phylogenetic matrix analysed and, if
  \code{run.now} is \code{TRUE}, a \code{multiPhylo} object of trees found
  from the search commands.
}
\description{
Phylogenetic analysis using implicit enumeration
}
