% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedWeights.R
\name{implied.weighting}
\alias{implied.weighting}
\title{Run an analysus using implied weighting}
\usage{
implied.weighting(tnt.path, analysis, k = 3, multi.k = FALSE)
}
\arguments{
\item{tnt.path}{the location of the TNT command-line binary.}

\item{analysis}{the parameters for an analysis as specified from the output
of the \code{implicit.enum}, \code{branchswap}, \code{ratchet}, or
\code{driven} commands.}

\item{k}{an integer value indicating the concavity constant to apply.}

\item{multi.k}{a logical value indicating whether each character will be
given an independent concavity constant based on the value of \code{k}.}
}
\description{
Run an analysus using implied weighting
}
