% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroImpliedWeights-class.R
\name{ImpliedWeightingOptions}
\alias{ImpliedWeightingOptions}
\title{Set options for implied weighting}
\description{
\code{ImpliedWeightingOptions} is an R6 class that defines the set of options
  for using (extended) implied weighting with phylogenetic analyses in
  \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{k}}{A numeric value indicating the concavity constant to apply
during an implied weights analysis. When \code{multi_k} is
\code{TRUE}, \code{k} represents the concavity for a character with
no missing entries and serves as the basis for calculating per-site
concavity constants.}

\item{\code{multi_k}}{A logical value indicating whether each character will be
given an independent concavity constant based on the value of \code{k}
during an implied weights analysis.}

\item{\code{proportion}}{A numeric value indicating the proportion of homoplasy
missing values are assumed to have under implied weighting. This
parameter is only valid when \code{multi_k} is \code{TRUE}. A
proportion of 0 is equivalent to standard implied weighting, with
\code{multi_k} set as \code{FALSE}.}

\item{\code{max_ratio}}{A numeric value indicating the maximum acceptable ratio
between two k values when multiple concavity constants are used (i.e.,
when \code{multi_k} is \code{TRUE}). A maximum ratio of 1 is
equivalent to standard implied weighting, with \code{multi_k} set as
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImpliedWeightingOptions-new}{\code{ImpliedWeightingOptions$new()}}
\item \href{#method-ImpliedWeightingOptions-print}{\code{ImpliedWeightingOptions$print()}}
\item \href{#method-ImpliedWeightingOptions-queue}{\code{ImpliedWeightingOptions$queue()}}
\item \href{#method-ImpliedWeightingOptions-clone}{\code{ImpliedWeightingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImpliedWeightingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImpliedWeightingOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImpliedWeightingOptions$new(
  k = 3,
  multi_k = FALSE,
  proportion = 0.5,
  max_ratio = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{A numeric value indicating the concavity constant to apply during
an implied weights analysis. When \code{multi_k} is \code{TRUE}, \code{k}
represents the concavity for a character with no missing entries and
serves as the basis for calculating per-site concavity constants.}

\item{\code{multi_k}}{A logical value indicating whether each character will be
given an independent concavity constant based on the value of \code{k}
during an implied weights analysis.}

\item{\code{proportion}}{A numeric value indicating the proportion of homoplasy
missing values are assumed to have under implied weighting. This parameter
is only valid when \code{multi_k} is \code{TRUE}. A proportion of 0 is
equivalent to standard implied weighting, with \code{multi_k} set as
\code{FALSE}.}

\item{\code{max_ratio}}{A numeric value indicating the maximum acceptable ratio
between two k values when multiple concavity constants are used (i.e.,
when \code{multi_k} is \code{TRUE}). A maximum ratio of 1 is equivalent
to standard implied weighting, with \code{multi_k} set as \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImpliedWeightingOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-ImpliedWeightingOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImpliedWeightingOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImpliedWeightingOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-ImpliedWeightingOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImpliedWeightingOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImpliedWeightingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImpliedWeightingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImpliedWeightingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
