% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{inactive.taxa}
\alias{inactive.taxa}
\title{Define inactive taxa}
\usage{
inactive.taxa(matrix, taxa)
}
\arguments{
\item{matrix}{a \code{phyDat} object of USER type.}

\item{taxa}{a character vector of taxa to be set as inactive.}
}
\value{
a \code{phyDat} object.
}
\description{
Define inactive taxa
}
