% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{inactiveCharacters}
\alias{inactiveCharacters}
\title{Define inactive characters}
\usage{
inactiveCharacters(matrix, characters)
}
\arguments{
\item{matrix}{a \code{phyDat} object of USER type.}

\item{characters}{a numeric vector of characters to be set as inactive.}
}
\value{
a \code{phyDat} object.
}
\description{
Define inactive characters
}
