% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroRatchet-methods.R
\name{iterations}
\alias{iterations}
\alias{iterations,NitroRatchet,missing-method}
\alias{iterations<-}
\alias{iterations<-,NitroRatchet,numeric-method}
\title{Iterations}
\usage{
iterations(n, value)

\S4method{iterations}{NitroRatchet,missing}(n)

iterations(n) <- value

\S4method{iterations}{NitroRatchet,numeric}(n) <- value
}
\arguments{
\item{n}{an object of either \code{NitroBranchSwap} or \code{NitroDriven}.}

\item{value}{a numeric vector indicating the number of iterations for the
analysis.}
}
\value{
a numeric vector indicating the number of iterations for the
analysis.
}
\description{
Function to return or set the number of iterations for a parsimony ratchet
phylogenetic analysis.
}
