% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintsBase-methods.R
\name{length,NitroConstraintsBase-method}
\alias{length,NitroConstraintsBase-method}
\title{Number of constraints}
\usage{
\S4method{length}{NitroConstraintsBase}(x)
}
\arguments{
\item{x}{a \code{"\linkS4class{NitroConstraintsBase}"} object}
}
\value{
an integer indicating the number of
  \code{"\linkS4class{NitroConstraint}"} objects.
}
\description{
Function that returns the number of constraints, in the form of
\code{"\linkS4class{NitroConstraint}"} objects, contained in a
\code{"\linkS4class{NitroConstraintsBase}"} object.
}
