% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MemoryAllocationCommand-class.R
\name{MemoryAllocationCommand}
\alias{MemoryAllocationCommand}
\title{Memory Allocation Command}
\description{
An \link[R6:R6Class]{R6} class that configures the amount of memory allocated
to TNT during execution.

This command is created automatically as part of the standard
\link{CommandQueue} setup. Users do not typically need to instantiate it
directly.
}
\details{
\subsection{Command output}{

\verb{$render()} produces a string of the form \verb{mxram \{size\};} where \code{{size}}
is the number of megabytes to allocate.
}

\subsection{Queue integration}{

Calling \verb{$enqueue()} adds this command to a \link{CommandQueue} at priority
\code{140}.
}
}
\keyword{internal}
\section{Super class}{
\code{\link[nitro:BasicCommand]{BasicCommand}} -> \code{MemoryAllocationCommand}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{size}}{[\code{integer(1)}]\cr
The amount of memory to allocate, in megabytes. Must be a
positive integer.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MemoryAllocationCommand-enqueue}{\code{MemoryAllocationCommand$enqueue()}}
    \item \href{#method-MemoryAllocationCommand-initialize}{\code{MemoryAllocationCommand$new()}}
    \item \href{#method-MemoryAllocationCommand-render}{\code{MemoryAllocationCommand$render()}}
    \item \href{#method-MemoryAllocationCommand-clone}{\code{MemoryAllocationCommand$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="format"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-format'><code>BasicCommand$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="get_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-get_dependency'><code>BasicCommand$get_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="new_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-new_dependency'><code>BasicCommand$new_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="print"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-print'><code>BasicCommand$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="set_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-set_dependency'><code>BasicCommand$set_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="transform"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-transform'><code>BasicCommand$transform()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryAllocationCommand-enqueue"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryAllocationCommand-enqueue}{}}}
\subsection{\code{MemoryAllocationCommand$enqueue()}}{
  Add this command to a \link{CommandQueue}.

Adds this command at priority \code{140} by default.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MemoryAllocationCommand$enqueue(.queue, priority = 140)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{.queue}}{A \link{CommandQueue} object}
      \item{\code{priority}}{[\code{integer(1)}]\cr
A non-negative integer controlling execution order. Lower values
execute first.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryAllocationCommand-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryAllocationCommand-initialize}{}}}
\subsection{\code{MemoryAllocationCommand$new()}}{
  Create a new \code{MemoryAllocateCommand} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MemoryAllocationCommand$new(size)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{size}}{[\code{integer(1)}]\cr
The amount of memory to allocate in megabytes. See the \verb{$size}
field.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{MemoryAllocateCommand} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryAllocationCommand-render"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryAllocationCommand-render}{}}}
\subsection{\code{MemoryAllocationCommand$render()}}{
  Render the TNT \code{mxram} command string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MemoryAllocationCommand$render(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Not used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A single-element character vector containing the TNT
command.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryAllocationCommand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryAllocationCommand-clone}{}}}
\subsection{\code{MemoryAllocationCommand$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MemoryAllocationCommand$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
