% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonophylyConstraintOptions-class.R
\name{MonophylyConstraintOptions}
\alias{MonophylyConstraintOptions}
\title{Set constraints on monophyly}
\description{
\code{MonophylyConstraintOptions} is an R6 class that sets options for tree
  searches with constraints on monophyly.
}
\section{Super class}{
\code{\link[nitro:ConstraintBaseOptions]{nitro::ConstraintBaseOptions}} -> \code{MonophylyConstraintOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fixed_otus}}{A logical vector indicating which OTUs from the matrix
to assign as a fixed constraint.}

\item{\code{floating_otus}}{An optional logical vector indicating which OTUs from
the matrix to assign as floating constraints.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MonophylyConstraintOptions-new}{\code{MonophylyConstraintOptions$new()}}
\item \href{#method-MonophylyConstraintOptions-print}{\code{MonophylyConstraintOptions$print()}}
\item \href{#method-MonophylyConstraintOptions-queue}{\code{MonophylyConstraintOptions$queue()}}
\item \href{#method-MonophylyConstraintOptions-clone}{\code{MonophylyConstraintOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonophylyConstraintOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-MonophylyConstraintOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonophylyConstraintOptions$new(
  fixed_otus,
  floating_otus = NULL,
  is_positive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fixed_otus}}{A character vector indicating which OTUs from the matrix
to assign as a fixed constraint.}

\item{\code{floating_otus}}{An character logical vector indicating which OTUs from
the matrix to assign as floating constraints.}

\item{\code{is_positive}}{A logical value indicating whether the constraint is
positive (\code{TRUE}) or negative (\code{FALSE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonophylyConstraintOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-MonophylyConstraintOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonophylyConstraintOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonophylyConstraintOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-MonophylyConstraintOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonophylyConstraintOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MonophylyConstraintOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MonophylyConstraintOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MonophylyConstraintOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
