% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_search_config.R
\name{new_search_config}
\alias{new_search_config}
\title{Create a new search configuration}
\usage{
new_search_config(name, ...)
}
\arguments{
\item{name}{The name of a search method. Partial unambiguous matches are also
accepted. The options are:
\itemize{
  \item \code{branch_breaking}: branch swapping on a set of trees. Creates
    a \code{\link{BranchBreakingOptions}} object;
  \item \code{branch_swapping}: branch swapping from a starting
    tree. Creates a \code{\link{BranchSwappingOptions}} object;
  \item \code{driven_search}: search using targets for completion. Creates
    a \code{\link{DrivenSearchOptions}} object;
  \item \code{ratchet}: search using parsimony ratchet. Creates a
    \code{\link{RatchetOptions}} object.
}}

\item{...}{Arguments to be passed on to the R6 class responsible for creating
the search method object.}
}
\description{
Create a new search configuration object for use in a tree analysis.
}
\seealso{
The \code{new} method of \code{\link{BranchBreakingOptions}},
  \code{\link{BranchSwappingOptions}}, \code{\link{DrivenSearchOptions}} and
  \code{\link{RatchetOptions}} describe optional arguments for search method
  creation.
}
