% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintsBase-methods.R
\name{newConstraint<-}
\alias{newConstraint<-}
\alias{newConstraint<-,NitroConstraintsBase,NitroConstraint-method}
\title{Add constraint}
\usage{
newConstraint(n) <- value

\S4method{newConstraint}{NitroConstraintsBase,NitroConstraint}(n) <- value
}
\arguments{
\item{n}{an object of class \code{"\linkS4class{NitroConstraintsBase}"}}

\item{value}{an object that inherits from
\code{"\linkS4class{NitroConstraint}"}}
}
\description{
Function to add a new \code{"\linkS4class{NitroConstraint}"} to a
\code{"\linkS4class{NitroConstraintsBase}"} object.
}
