% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{newTreeSearch}
\alias{newTreeSearch}
\title{Define tree search parameters}
\usage{
newTreeSearch(
  matrix,
  method,
  ordered_characters = numeric(),
  inactive_taxa = character(),
  inactive_characters = numeric(),
  outgroup = NULL,
  collapse = 3,
  weights = NULL,
  start_trees = NULL,
  combine = FALSE,
  hold = 100,
  max_ram = 16
)
}
\arguments{
\item{matrix}{a \code{phyDat} object representing a character-taxon matrix}

\item{method}{an object that inherits from class
\code{"\linkS4class{NitroMethodsBase}"}}

\item{ordered_characters}{a numeric vector indicating the characters to be treated as ordered.}

\item{inactive_taxa}{a character vector indicating the taxa to be marked as inactive.}

\item{inactive_characters}{a numeric vector indicating the characters to be marked as inactive.}

\item{outgroup}{a single character vector indicating the taxon to be the outgroup.}

\item{collapse}{an integer indicating the rule for collapsing of zero length branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}

\item{weights}{an object inheriting class}

\item{start_trees}{an list of class \code{multiPhylo} containing trees to
start the analysis with. Required for analyses using
\code{"\linkS4class{NitroRatchet}"}.}

\item{combine}{a logical indicating whether to merge any start trees from
\code{obj} with the trees obtained from the analysis, removing any
duplicates. This option is useful when a large number of trees, or any
number of trees with a large number of tips, from two separate analyses
(i.e., such as when a set of most parsimonious trees are further explored
with branch breaking) are to be combined, excluding duplicates. TNT's
method for identifying duplicate trees is significantly more efficient
than current R implementations (e.g., \code{\link{unique.phylo}}).}

\item{hold}{an integer indicating the number of trees to hold in TNTs tree
buffer.}

\item{max_ram}{a numeric indicating the number of (binary) megabytes to
allocate for use by TNT.
\code{"\linkS4class{NitroWeightsBase}"}.}
}
\value{
an object of class \code{"\linkS4class{NitroTreeSearch}"}.
}
\description{
Define tree search parameters
}
