% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBase-class.R
\docType{class}
\name{NitroBase-class}
\alias{NitroBase-class}
\title{Define base parameters}
\description{
\code{NitroBase} is a virtual S4 class that defines the basic set of
parameters required to perform a phylogenetic analysis in \code{nitro}.
As this is a virtual class it cannot be constructed directly.
}
\seealso{
The S4 classes \code{NitroTreeSearch}, \code{NitroEqualWeights}
and \code{NitroImpliedWeights}.
}
\keyword{classes}
