% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBootstrap-class.R
\docType{class}
\name{NitroBootstrap-class}
\alias{NitroBootstrap-class}
\title{Define bootstrap resampling analysis}
\description{
\code{NitroBootstrap} is an S4 class that defines parameters for
bootstrap resampling analyses.
}
\seealso{
The inheriting S4 class
\code{"\linkS4class{NitroConstraintsBase}"}.
}
\keyword{classes}
