% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBootstrap-methods.R
\name{NitroBootstrap}
\alias{NitroBootstrap}
\title{Perform bootstrap resampling}
\usage{
NitroBootstrap(cutoff = 0, tree_search = NULL, phy = NULL, replications = 100)
}
\arguments{
\item{cutoff}{an integer value indicating the cutoff value for
frequencies.}

\item{tree_search}{an object inheriting class
\code{"\linkS4class{NitroMethodsBase}"}.}

\item{phy}{a tree of class \code{phylo}. Resampling values will be
calculated using this topology.}

\item{replications}{an integer value indicating the number of resampling
replications to perform.}
}
\description{
Perform bootstrap resampling
}
