% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchBreak-class.R
\name{NitroBranchBreak}
\alias{NitroBranchBreak}
\title{Define parameters for branch breaking}
\description{
\code{NitroBranchBreak} is an R6 class that defines the set of parameters
required to perform branch breaking.
}
\section{Super class}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{NitroBranchBreak}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{swapper}}{A character vector indicating the branch swapping algorithm
to use. Options are either '\code{tbr}' (the default) or '\code{spr}'.}

\item{\code{cluster_size}}{An integer value indicating the number of nodes
(clusters) to use during swapping. Applied only when \code{swapper} is
set to \code{tbr}. Larger clusters will result in faster swapping as
matrix size increases. If left as 0 (the default), cluster size will be
automatically determined.}

\item{\code{safe_unclip}}{A logical value indicating whether to use a safer but
slower method for updating buffers when clipping to find a better tree.
Applied only when \code{swapper} is set to \code{tbr}.}

\item{\code{fill_only}}{A logical value indicating whether to stop swapping when
the tree buffer is full.}

\item{\code{save_multiple}}{A logical value indicating whether to save multiple
trees during swapping.}

\item{\code{random_clip}}{A logical value indicating whether to randomize the
tree clipping sequence.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroBranchBreak$new()}}
\item \href{#method-print}{\code{NitroBranchBreak$print()}}
\item \href{#method-tnt_cmd}{\code{NitroBranchBreak$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroBranchBreak$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroBranchBreak$new(
  swapper = "tbr",
  cluster_size = 0,
  safe_unclip = FALSE,
  fill_only = FALSE,
  save_multiple = TRUE,
  random_clip = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{swapper}}{A character vector indicating the branch swapping algorithm
to use. Options are either '\code{tbr}' (the default) or '\code{spr}'.}

\item{\code{cluster_size}}{An integer value indicating the number of nodes
(clusters) to use during swapping. Applied only when \code{swapper} is
set to \code{tbr}. Larger clusters will result in faster swapping as
matrix size increases. If left as 0 (the default), cluster size will be
automatically determined.}

\item{\code{safe_unclip}}{A logical value indicating whether to use a safer but
slower method for updating buffers when clipping to find a better tree.
Applied only when \code{swapper} is set to \code{tbr}.}

\item{\code{fill_only}}{A logical value indicating whether to stop swapping when
the tree buffer is full.}

\item{\code{save_multiple}}{A logical value indicating whether to save multiple
trees during swapping.}

\item{\code{random_clip}}{A logical value indicating whether to randomize the
tree clipping sequence.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroBranchBreak$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroBranchBreak$tnt_cmd(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroBranchBreak$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
