% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchSwap-class.R
\docType{class}
\name{NitroBranchSwap-class}
\alias{NitroBranchSwap-class}
\title{Define a branch swapping analysis}
\description{
\code{NitroBranchSwap} is an S4 class that defines the set of parameters
required to perform a branch swapping ('traditional', in TNTs terminology)
phylogenetic analysis in \code{nitro}.
}
\seealso{
The S4 classes \code{ImplicitEnum}, \code{BranchSwap}, \code{Ratchet}
and \code{Driven}.
}
\keyword{classes}
