% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraint-class.R
\docType{class}
\name{NitroConstraint-class}
\alias{NitroConstraint-class}
\title{Define constrained tree search}
\description{
\code{NitroConstraint} is an S4 class that defines a constrainted tree
search.
}
\seealso{
The inheriting S4 class \code{\link{NitroConstraintsBase}} and
\code{\link{NitroUnconstrainedSearch}}.
}
\keyword{classes}
