% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintsBase-class.R
\docType{class}
\name{NitroConstraintsBase-class}
\alias{NitroConstraintsBase-class}
\title{Define constraints on monophyly}
\description{
\code{NitroConstraintsBase} is an S4 class that serves as the basis for
classes that define parameters for constraints on monophyly.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroConstraint}"}.
}
\keyword{classes}
