% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintSectorialSearch-class.R
\name{NitroConstraintSectorialSearch}
\alias{NitroConstraintSectorialSearch}
\title{Define a constraint sectorial search analysis}
\description{
\code{NitroConstraintSectorialSearch} is an R6 class that defines the set of
parameters required to perform sectorial searching analyses in
\code{nitro}.
}
\section{Super classes}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{\link[nitro:NitroSectorialSearch]{nitro::NitroSectorialSearch}} -> \code{NitroConstraintSectorialSearch}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{min_fork}}{An integer value indicating the minimum fork number to use
with constraint-based sectorial searches.}

\item{\code{max_fork}}{An integer value indicating the maximum fork number to use
with constraint-based sectorial searches.}

\item{\code{rounds}}{An integer value indicating the number of times to cycle over
groups in constraint-based selections.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroConstraintSectorialSearch$new()}}
\item \href{#method-print}{\code{NitroConstraintSectorialSearch$print()}}
\item \href{#method-tnt_cmd}{\code{NitroConstraintSectorialSearch$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroConstraintSectorialSearch$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroConstraintSectorialSearch$new(
  min_fork = 10,
  max_fork = 10,
  rounds = 3,
  buffer = TRUE,
  slack = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_fork}}{An integer value indicating the minimum fork number to use
with constraint-based sectorial searches.}

\item{\code{max_fork}}{An integer value indicating the maximum fork number to use
with constraint-based sectorial searches.}

\item{\code{rounds}}{An integer value indicating the number of times to cycle over
groups in constraint-based selections.}

\item{\code{buffer}}{A logical value indicating whether to use an independent
memory buffer for analysis of sectors.}

\item{\code{slack}}{An integer value indicating the percentage to increase the
available memory during searches.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroConstraintSectorialSearch$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroConstraintSectorialSearch$tnt_cmd(set_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_only}}{A logical indicating whether to instruct the command to
execute immediately (\code{TRUE}) or set the variables for future
execution \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroConstraintSectorialSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
