% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroDriven-methods.R
\name{NitroDriven}
\alias{NitroDriven}
\title{Definte parameters for a driven analysis}
\usage{
NitroDriven(
  replications,
  hits = 1,
  rss = TRUE,
  css = FALSE,
  xss = FALSE,
  ratchet_cycles = 0,
  drifting_cycles = 30,
  fusing_rounds = 0,
  consense_times = 0,
  keep_all = FALSE,
)
}
\arguments{
\item{replications}{an integer value indicating the number of replications.}

\item{hits}{an integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{rss, css, xss}{a logical value indicating whether to use random,
constraint or exclusive sectorial searches.}

\item{ratchet_cycles}{an integer value indicating the number of cycles of
the parsimony ratchet to perform per replication.}

\item{drifting_cycles}{an integer value indicating the number of cycles of
tree drifting to perform per replication.}

\item{fusing_rounds}{an integer value indicating the number of rounds of
fusing per replication.}

\item{consense_times}{an integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{keep_all}{a logical value indicating whether to retain all generated
trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When
\code{hits} = 1, it is trees from each of the RAS + TBR +  SS or DFT or
RAT, in addition to the trees resulting from fusing those. When
\code{hits} > 1, then it means the trees resulting from fusing the
initial starting trees for each of starting points.}






}
\description{
Definte parameters for a driven analysis
}
