% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroEqualWeights-class.R
\docType{class}
\name{NitroEqualWeights-class}
\alias{NitroEqualWeights-class}
\title{Define equal weights analysis}
\description{
\code{NitroEqualWeights} is an S4 class that serves as the basis for classes
that define parameters for equal weights phylogenetic analyses.
}
\seealso{
The inheriting S4 classes \code{NitroImplicitEnum},
\code{NitroBranchswap}, \code{NitroRatchet}, \code{NitroDriven}.
}
\keyword{classes}
