% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroEqualWeights-class.R
\name{NitroEqualWeights}
\alias{NitroEqualWeights}
\title{Define equal weights analysis}
\description{
\code{NitroEqualWeights} is an R6 class that serves as the basis for classes
that define parameters for equal weights phylogenetic analyses.
}
\section{Super class}{
\code{\link[nitro:NitroWeightsBase]{nitro::NitroWeightsBase}} -> \code{NitroEqualWeights}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{NitroEqualWeights$print()}}
\item \href{#method-clone}{\code{NitroEqualWeights$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroEqualWeights$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroEqualWeights$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
