% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroImpliedWeights-class.R
\docType{class}
\name{NitroImpliedWeights-class}
\alias{NitroImpliedWeights-class}
\title{Define implied weights analysis}
\description{
\code{NitroImpliedWeights} is an S4 class that serves as the basis for
classes that define parameters for implied weights phylogenetic analyses.
}
\seealso{
The inheriting S4 class \code{"\linkS4class{NitroWeightsBase}"} and
\code{"\linkS4class{NitroEqualWeights}"}.
}
\keyword{classes}
