% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroImpliedWeights-class.R
\name{NitroImpliedWeights}
\alias{NitroImpliedWeights}
\title{Define implied weights analysis}
\description{
\code{NitroImpliedWeights} is an R6 class that serves as the basis for
classes that define parameters for implied weights phylogenetic analyses.
}
\section{Super class}{
\code{\link[nitro:NitroWeightsBase]{nitro::NitroWeightsBase}} -> \code{NitroImpliedWeights}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{k}}{A numeric value indicating the concavity constant to apply
during an implied weights analysis. When \code{multi_k} is
\code{TRUE}, \code{k} represents the concavity for a character with
no missing entries and serves as the basis for calculating per-site
concavity constants.}

\item{\code{multi_k}}{A logical value indicating whether each character will be
given an independent concavity constant based on the value of \code{k}
during an implied weights analysis.}

\item{\code{proportion}}{A numeric value indicating the proportion of homoplasy
missing values are assumed to have under implied weighting. This
parameter is only valid when \code{multi_k} is \code{TRUE}. A
proportion of 0 is equivalent to standard implied weighting, with
\code{multi_k} set as \code{FALSE}.}

\item{\code{max_ratio}}{A numeric value indicating the maximum acceptable ratio
between two k values when multiple concavity constants are used (i.e.,
when \code{multi_k} is \code{TRUE}). A maximum ratio of 1 is
equivalent to standard implied weighting, with \code{multi_k} set as
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroImpliedWeights$new()}}
\item \href{#method-print}{\code{NitroImpliedWeights$print()}}
\item \href{#method-tnt_cmd}{\code{NitroImpliedWeights$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroImpliedWeights$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroImpliedWeights$new(
  k = 3,
  multi_k = FALSE,
  proportion = 0.5,
  max_ratio = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{A numeric value indicating the concavity constant to apply during
an implied weights analysis. When \code{multi_k} is \code{TRUE}, \code{k}
represents the concavity for a character with no missing entries and
serves as the basis for calculating per-site concavity constants.}

\item{\code{multi_k}}{A logical value indicating whether each character will be
given an independent concavity constant based on the value of \code{k}
during an implied weights analysis.}

\item{\code{proportion}}{A numeric value indicating the proportion of homoplasy
missing values are assumed to have under implied weighting. This parameter
is only valid when \code{multi_k} is \code{TRUE}. A proportion of 0 is
equivalent to standard implied weighting, with \code{multi_k} set as
\code{FALSE}.}

\item{\code{max_ratio}}{A numeric value indicating the maximum acceptable ratio
between two k values when multiple concavity constants are used (i.e.,
when \code{multi_k} is \code{TRUE}). A maximum ratio of 1 is equivalent
to standard implied weighting, with \code{multi_k} set as \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroImpliedWeights$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroImpliedWeights$tnt_cmd(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroImpliedWeights$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
