% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroJackknife-class.R
\docType{class}
\name{NitroJackknife-class}
\alias{NitroJackknife-class}
\title{Define jackknife resampling analysis}
\description{
\code{NitroJackknife} is an S4 class that defines parameters for
jackknife resampling analyses.
}
\seealso{
The inheriting S4 class
\code{"\linkS4class{NitroConstraintsBase}"}.
}
\keyword{classes}
