% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroJacknife-class.R
\docType{class}
\name{NitroJacknife-class}
\alias{NitroJacknife-class}
\title{Define jacknife resampling analysis}
\description{
\code{NitroJacknife} is an S4 class that defines parameters for
jacknife resampling analyses.
}
\seealso{
The inheriting S4 class
\code{"\linkS4class{NitroConstraintsBase}"}.
}
\keyword{classes}
