% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroJacknife-methods.R
\name{NitroJacknife}
\alias{NitroJacknife}
\title{Perform jacknife resampling}
\usage{
NitroJacknife(
  probability = 36,
  cutoff = 0,
  tree_search = NULL,
  phy = NULL,
  replications = 100
)
}
\arguments{
\item{probability}{an integer value indicating the change probability.}

\item{cutoff}{an integer value indicating the cutoff value for
frequencies.}

\item{tree_search}{an object inheriting class
\code{"\linkS4class{NitroMethodsBase}"}.}

\item{phy}{a tree of class \code{phylo}. Resampling values will be
calculated using this topology.}

\item{replications}{an integer value indicating the number of resampling
replications to perform.}
}
\description{
Perform jacknife resampling
}
