% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroMethodsBase-class.R
\docType{class}
\name{NitroMethodsBase-class}
\alias{NitroMethodsBase-class}
\title{Define tree searches}
\description{
\code{NitroMethodsBase} is an S4 class that serves as the basis for classes that
define parameters for tree search methods.
}
\seealso{
The inheriting S4 classes \code{"\linkS4class{NitroImplicitEnum}"},
\code{"\linkS4class{NitroBranchSwap}"},
\code{"\linkS4class{NitroRatchet}"} and
\code{"\linkS4class{NitroDriven}"}.
}
\keyword{classes}
