% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroRandomSectorialSearch-class.R
\docType{class}
\name{NitroRandomSectorialSearch-class}
\alias{NitroRandomSectorialSearch-class}
\title{Define a random sectorial search analysis}
\description{
\code{NitroRandomSectorialSearch} is an S4 class that defines the set of
parameters required to perform sectorial searching analyses in
\code{nitro}.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroImplicitEnum}"},
\code{"\linkS4class{NitroRatchet}"} and \code{"\linkS4class{NitroDriven}"}.
}
\keyword{classes}
