% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroRandomSectorialSearch-methods.R
\name{NitroRandomSectorialSearch}
\alias{NitroRandomSectorialSearch}
\title{Define parameters for a random sectorial search analysis}
\usage{
NitroRandomSectorialSearch(
  min_size = 0,
  max_size = 0,
  selection_factor = 50,
  increase = 100,
  small_starts = 3,
  buffer = TRUE,
  slack = 0
)
}
\arguments{
\item{min_size}{an integer value indicating the minimum size of random
selections.}

\item{max_size}{an integer value indicating the maximum size of random
selections.}

\item{selection_factor}{an integer value indicating the maximum number of
random selections for the currently active taxa.}

\item{increase}{an integer value indicating the factor to increase the size
of random selections if enough selection of the current size have been
completed.}

\item{small_starts}{an integer value indicating the number of random
addition sequence plus tree-bisection reconnection swaps to perform for
random selections below \code{min_size}.}

\item{buffer}{a logical value indicating whether to use an independent}

\item{slack}{an integer value indicating the percentage to increase the
available memory during searches.}
}
\description{
Define parameters for a random sectorial search analysis
}
