% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroRatchet-class.R
\docType{class}
\name{NitroRatchet-class}
\alias{NitroRatchet-class}
\title{Define a parsimony ratchet analysis}
\description{
\code{NitroRatchet} is an S4 class that defines the set of parameters
required to perform a parsimony ratchet phylogenetic analysis in
\code{nitro}.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroImplicitEnum}"},
\code{"\linkS4class{NitroRatchet}"} and \code{"\linkS4class{NitroDriven}"}.
}
\keyword{classes}
