% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroRatchet-methods.R
\name{NitroRatchet}
\alias{NitroRatchet}
\title{Define parameters for a parsimony ratchet analysis}
\usage{
NitroRatchet(iterations = 50, replacements = 40, prob_up = 4, prob_down = 4)
}
\arguments{
\item{iterations}{an integer value indicating the number of iterations.}

\item{replacements}{an integer value indicating the number of replacements
(i.e., accepted tree rearrangements) to perform in each perturbation
phase.}

\item{prob_up}{an integer value indicating the probability of upweighting a
character.}

\item{prob_down}{an integer value indicating the probability of
downweighting a character.}
}
\description{
Define parameters for a parsimony ratchet analysis
}
