% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroResample-class.R
\docType{class}
\name{NitroResample-class}
\alias{NitroResample-class}
\title{Define resampling analysis}
\description{
\code{NitroResample} is an S4 class that defines a resampling tree search.
}
\seealso{
The inheriting S4 class \code{\link{NitroConstraintsBase}} and
\code{\link{NitroUnconstrainedSearch}}.
}
\keyword{classes}
