% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroResampleBase-class.R,
%   R/NitroJacknife-class.R
\docType{class}
\name{NitroResampleBase-class}
\alias{NitroResampleBase-class}
\title{Define resampling parameters}
\description{
\code{NitroResampleBase} is an S4 class that defines parameters for
resampling analyses.

\code{NitroResampleBase} is an S4 class that defines parameters for
jacknife resampling analyses.
}
\seealso{
The inheriting S4 class
\code{"\linkS4class{NitroConstraintsBase}"}.

The inheriting S4 class
\code{"\linkS4class{NitroConstraintsBase}"}.
}
\keyword{classes}
