% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroResampleBase-class.R
\name{NitroResampleBase}
\alias{NitroResampleBase}
\title{Define resampling parameters}
\description{
\code{NitroResampleBase} is an R6 class that defines parameters for
resampling analyses.
}
\section{Super class}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{NitroResampleBase}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{search_method}}{An object inheriting class
\code{"\link{NitroMethodsBase}"}.}

\item{\code{phy}}{A tree of class \code{phylo}. Resampling values will be
calculated using this topology.}

\item{\code{replications}}{An integer value indicating the number of resampling
replications to perform.}

\item{\code{abs_freq_summary}}{A logical value indicating whether to summarize
supports using absolute frequencies.}

\item{\code{freq_diff_summary}}{A logical value indicating whether to summarize
supports using frequency differences (i.e.,
Group supported/Contradicted).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroResampleBase$new()}}
\item \href{#method-clone}{\code{NitroResampleBase$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroResampleBase$new(search_method, phy, replications = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_method}}{An object inheriting class
\code{"\link{NitroMethodsBase}"}.}

\item{\code{phy}}{A tree of class \code{phylo}. Resampling values will be
calculated using this topology.}

\item{\code{replications}}{An integer value indicating the number of resampling
replications to perform.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroResampleBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
