% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroResults-class.R
\docType{class}
\name{NitroResults-class}
\alias{NitroResults-class}
\title{Define analysis results}
\description{
\code{NitroResults} is an S4 class that contains the results of an analysis
performed using \code{nitro}.
}
\keyword{classes}
