% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroSymmetricResample-class.R
\docType{class}
\name{NitroSymmetricResample-class}
\alias{NitroSymmetricResample-class}
\title{Define symmetric resampling analysis}
\description{
\code{NitroSymmetricResample} is an S4 class that defines parameters for
symmetric resampling analyses.
}
\seealso{
The inheriting S4 class
\code{"\linkS4class{NitroConstraintsBase}"}.
}
\keyword{classes}
