% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroSymmetricResample-class.R
\name{NitroSymmetricResample}
\alias{NitroSymmetricResample}
\title{Define symmetric resampling analysis}
\description{
\code{NitroSymmetricResample} is an R6 class that defines parameters for
symmetric resampling analyses.
}
\section{Super classes}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{\link[nitro:NitroResampleBase]{nitro::NitroResampleBase}} -> \code{NitroSymmetricResample}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{probability}}{An integer value indicating the change probability.}

\item{\code{cutoff}}{An integer value indicating the cutoff value for
frequencies.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroSymmetricResample$new()}}
\item \href{#method-print}{\code{NitroSymmetricResample$print()}}
\item \href{#method-tnt_cmd}{\code{NitroSymmetricResample$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroSymmetricResample$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroSymmetricResample$new(
  probability = 36,
  cutoff = 0,
  search_method = NULL,
  phy = NULL,
  replications = 100
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probability}}{An integer value indicating the change probability.}

\item{\code{cutoff}}{An integer value indicating the cutoff value for
frequencies.}

\item{\code{search_method}}{An object inheriting class
\code{"\link{NitroMethodsBase}"}.}

\item{\code{phy}}{A tree of class \code{phylo}. Resampling values will be
calculated using this topology.}

\item{\code{replications}}{an integer value indicating the number of resampling
replications to perform.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroSymmetricResample$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroSymmetricResample$tnt_cmd(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroSymmetricResample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
