% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeDrift-class.R
\docType{class}
\name{NitroTreeDrift-class}
\alias{NitroTreeDrift-class}
\title{Define tree drift properties}
\description{
\code{NitroTreeDrift} is an S4 class that defines the set of parameters
required for performing tree drift operations in \code{nitro}.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroImplicitEnum}"},
\code{"\linkS4class{NitroRatchet}"} and \code{"\linkS4class{NitroDriven}"}.
}
\keyword{classes}
