% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeDrift-methods.R
\name{NitroTreeDrift}
\alias{NitroTreeDrift}
\title{Define parameters for tree drift operations}
\usage{
NitroTreeDrift(
  iterations = 30,
  substitutions = 60,
  max_abs_fit_diff = 1,
  max_rel_fit_diff = 0.2,
  reject_factor = 3,
  autoconstrain_cycles = 0
)
}
\arguments{
\item{iterations}{an integer value indicating the number of tree-drifting
cycles to perform.}

\item{substitutions}{an integer value indicating the number of replacements
(i.e., accepted tree rearrangements) to perform in the perturbation phase.}

\item{max_abs_fit_diff}{a numeric value indicating the maximum absolute fit
difference.}

\item{max_rel_fit_diff}{a numeric value indicating the maximum relative fit
difference.}

\item{reject_factor}{a numeric value indicating the rejection factor for
suboptimal trees.}

\item{autoconstrain_cycles}{an integer value indicating the number of
autoconstrained cycles to perform.}
}
\description{
Define parameters for tree drift operations
}
