% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeFuse-class.R
\docType{class}
\name{NitroTreeFuse-class}
\alias{NitroTreeFuse-class}
\title{Define tree fusing properties}
\description{
\code{NitroTreeFuse} is an S4 class that defines the set of parameters
required for performing tree fusing operations in \code{nitro}.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroImplicitEnum}"},
\code{"\linkS4class{NitroRatchet}"} and \code{"\linkS4class{NitroDriven}"}.
}
\keyword{classes}
