% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeFuse-methods.R
\name{NitroTreeFuse}
\alias{NitroTreeFuse}
\title{Define parameters for tree fusing operations}
\usage{
NitroTreeFuse(
  rounds = 5,
  exchange_equal = FALSE,
  start_best = TRUE,
  keep_all = TRUE,
  accept_all = TRUE,
  swap = TRUE
)
}
\arguments{
\item{rounds}{an integer value indicating the number of tree-fusing rounds
to perform.}

\item{exchange_equal}{a logical value indicating whether to accept exchanges
of equal score.}

\item{start_best}{a logical value indicating whether to use the best tree to
start tree-fusing.}

\item{keep_all}{a logical value indicating whether to keep all trees found
instead of only the best trees.}

\item{accept_all}{a logical value indicating whether to accept all exchanges
rather than only those that improve the best score.}

\item{swap}{a logical value indicating whether to perform tree-bisection
reconnection swapping after exchanging clades.}
}
\description{
Define parameters for tree fusing operations
}
