% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeHybridize-class.R
\docType{class}
\name{NitroTreeHybridize-class}
\alias{NitroTreeHybridize-class}
\title{Define tree hybridization properties}
\description{
\code{NitroTreeHybridize} is an S4 class that defines the set of parameters
required for performing tree hybridizing operations in \code{nitro}.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroImplicitEnum}"},
\code{"\linkS4class{NitroRatchet}"} and \code{"\linkS4class{NitroDriven}"}.
}
\keyword{classes}
