% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeHybridize-methods.R
\name{NitroTreeHybridize}
\alias{NitroTreeHybridize}
\title{Define parameters for tree hybridizing operations}
\usage{
NitroTreeHybridize(
  rounds = 1,
  hybridizations = 1000,
  best_trees = 50,
  replace = TRUE,
  sample_factor = 15
)
}
\arguments{
\item{rounds}{an integer value indicating the number of rounds of
tree-hybridizing to perform.}

\item{hybridizations}{an integer value indicating the number of
hybridizations to perform in each round.}

\item{best_trees}{an integer value indicating the number of best trees from
the previous round of hybridizing to use in the next round.}

\item{replace}{a logical value indicating whether to replace the source tree
with a better tree produced by hybridizing.}

\item{sample_factor}{an integer value indicating the number of times to
increase the size of initial tree set by. The corresponding number of trees
to retain will be proportional to the inverse of this value.}
}
\description{
Define parameters for tree hybridizing operations
}
