% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTrees-class.R
\docType{class}
\name{NitroTrees-class}
\alias{NitroTrees-class}
\title{Define analysis results}
\description{
\code{NitroTrees} is an S4 class that contains the
\code{"\linkS4class{NitroTreeSearch}"} defining the parameters for an
analysis the set of trees that the analysis returned.
}
\keyword{classes}
