% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-class.R
\docType{class}
\name{NitroTreeSearch-class}
\alias{NitroTreeSearch-class}
\title{Define tree searches}
\description{
\code{NitroTreeSearch} is an S4 class that stores information regarding the
tree search method, character weighting scheme and constraints on monophyly.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroConstraintsBase}"},
\code{"\linkS4class{NitroMethodsBase}"} and
\code{"\linkS4class{NitroWeightsBase}"}.
}
\keyword{classes}
