% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-class.R
\name{NitroTreeSearch}
\alias{NitroTreeSearch}
\title{Define tree searches}
\description{
\code{NitroTreeSearch} is an R6 class that stores information regarding the
tree search method, character weighting scheme and constraints on monophyly.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{matrix}}{A \code{phyDat} object representing a character-taxon matrix}

\item{\code{method}}{An object that inherits from class
\code{"\link{NitroMethodsBase}"}}

\item{\code{ordered_characters}}{A numeric vector indicating the characters to
be treated as ordered.}

\item{\code{inactive_taxa}}{A character vector indicating the taxa to be
inactivated.}

\item{\code{inactive_characters}}{A numeric vector indicating the characters to
be inactivated.}

\item{\code{outgroup}}{A single character vector indicating the taxon to be
the outgroup.}

\item{\code{collapse}}{An integer indicating the rule for collapsing of zero
length branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible
  length of the branch is zero;
\item \code{2}: keep zero length branches if ancestor and descendant
  states differ;
\item \code{3}: collapse an interior branch if the minimum possible
  length of the branch is zero (the default); and
\item \code{4}: discard all trees that must contain a zero length
  branch.
}}

\item{\code{constraints}}{A list of \code{"\link{NitroConstraint}"} objects.}

\item{\code{weights}}{An object inheriting from
\code{"\link{NitroWeightsBase}"}.}

\item{\code{start_trees}}{A \code{phylo} or \code{multiPhylo} of trees to start
the analysis with. Required for analyses using
\code{"\link{NitroRatchet}"}.}

\item{\code{combine}}{A logical indicating whether to merge any start trees
from \code{obj} with the trees obtained from the analysis, removing
any duplicates. This option is useful when a large number of trees, or
any number of trees with a large number of tips, from two separate
analyses (i.e., such as when a set of most parsimonious trees are
further explored with branch breaking) are to be combined, excluding
duplicates. TNT's method for identifying duplicate trees is
significantly more efficient than current R implementations (e.g.,
\code{\link[ape:unique.multiPhylo]{ape::unique.multiPhylo()}}).}

\item{\code{hold}}{An integer indicating the number of trees to hold in TNT's
tree buffer.}

\item{\code{max_ram}}{A numeric indicating the number of (binary) megabytes to
allocate for use by TNT.
\code{"\link{NitroWeightsBase}"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroTreeSearch$new()}}
\item \href{#method-print}{\code{NitroTreeSearch$print()}}
\item \href{#method-add_constraint}{\code{NitroTreeSearch$add_constraint()}}
\item \href{#method-clone}{\code{NitroTreeSearch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeSearch$new(
  matrix,
  method,
  ordered_characters = NULL,
  inactive_taxa = NULL,
  inactive_characters = NULL,
  outgroup = NULL,
  collapse = 3,
  weights = NULL,
  start_trees = NULL,
  combine = FALSE,
  hold = 100,
  max_ram = 16
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{A \code{phyDat} object representing a character-taxon matrix}

\item{\code{method}}{An object that inherits from class
\code{"\link{NitroMethodsBase}"}}

\item{\code{ordered_characters}}{A numeric vector indicating the characters to
be treated as ordered.}

\item{\code{inactive_taxa}}{A character vector indicating the taxa to be
inactivated.}

\item{\code{inactive_characters}}{A numeric vector indicating the characters to
be inactivated.}

\item{\code{outgroup}}{A single character vector indicating the taxon to be the
outgroup.}

\item{\code{collapse}}{An integer indicating the rule for collapsing of zero
length branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible
  length of the branch is zero;
\item \code{2}: keep zero length branches if ancestor and descendant
  states differ;
\item \code{3}: collapse an interior branch if the minimum possible
  length of the branch is zero (the default); and
\item \code{4}: discard all trees that must contain a zero length
  branch.
}}

\item{\code{weights}}{An object inheriting from \code{"\link{NitroWeightsBase}"}.}

\item{\code{start_trees}}{A \code{phylo} or \code{multiPhylo} of trees to start
the analysis with. Required for analyses using
\code{"\link{NitroRatchet}"}.}

\item{\code{combine}}{A logical indicating whether to merge any start trees
from \code{obj} with the trees obtained from the analysis, removing
any duplicates. This option is useful when a large number of trees, or
any number of trees with a large number of tips, from two separate
analyses (i.e., such as when a set of most parsimonious trees are
further explored with branch breaking) are to be combined, excluding
duplicates. TNT's method for identifying duplicate trees is
significantly more efficient than current R implementations (e.g.,
\code{\link[ape:unique.multiPhylo]{ape::unique.multiPhylo()}}).}

\item{\code{hold}}{An integer indicating the number of trees to hold in TNTs tree
buffer.}

\item{\code{max_ram}}{A numeric indicating the number of (binary) megabytes to
allocate for use by TNT.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeSearch$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.
Add new constraint

Function to add a new constraint to a constraint group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-add_constraint}{}}}
\subsection{Method \code{add_constraint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeSearch$add_constraint(
  fixed_taxa,
  floating_taxa = character(),
  type = c("positive", "negative")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fixed_taxa}}{a character vector of taxa to set as fixed constraints}

\item{\code{floating_taxa}}{a character vector of taxa to set as floating
constraints}

\item{\code{type}}{a character vector indicating whether the constraint will be
'\code{positive}' or '\code{negative}'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
