% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroWeightsBase-class.R
\docType{class}
\name{NitroWeightsBase-class}
\alias{NitroWeightsBase-class}
\title{Define tree searches}
\description{
\code{NitroWeighBase} is an S4 class that serves as the basis for classes
that define parameters for equal and implied weights analyses.
}
\seealso{
The inheriting S4 classes \code{"\linkS4class{NitroEqualWeights}"}
and \code{"\linkS4class{NitroImpliedWeights}"}.
}
\keyword{classes}
