% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{orderedCharacters}
\alias{orderedCharacters}
\title{Define ordered characters}
\usage{
orderedCharacters(matrix, characters)
}
\arguments{
\item{matrix}{a \code{phyDat} object of USER type.}

\item{characters}{a vector of characters to be set as ordered.}
}
\value{
a \code{phyDat} object.
}
\description{
Define ordered characters
}
