% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBase-methods.R
\name{outgroup}
\alias{outgroup}
\alias{outgroup,NitroBase-method}
\alias{outgroup<-}
\alias{outgroup<-,NitroBase,character-method}
\title{Outgroup}
\usage{
outgroup(n)

\S4method{outgroup}{NitroBase}(n)

outgroup(n) <- value

\S4method{outgroup}{NitroBase,character}(n) <- value
}
\arguments{
\item{n}{an object that inherits \code{NitroBase}.}

\item{value}{a single character vector indicating the taxon to be the outgroup.}
}
\value{
a character vector indicating the name of the outgroup taxon.
}
\description{
Function to return or set the outgroup taxon.
}
