% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputParser-class.R
\name{OutputParser}
\alias{OutputParser}
\title{Output parser for TNT}
\description{
\code{OutputParser} is an R6 class that parses output from TNT.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OutputParser-new}{\code{OutputParser$new()}}
\item \href{#method-OutputParser-clean}{\code{OutputParser$clean()}}
\item \href{#method-OutputParser-content_detect}{\code{OutputParser$content_detect()}}
\item \href{#method-OutputParser-transform}{\code{OutputParser$transform()}}
\item \href{#method-OutputParser-clone}{\code{OutputParser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputParser-new"></a>}}
\if{latex}{\out{\hypertarget{method-OutputParser-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputParser$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputParser-clean"></a>}}
\if{latex}{\out{\hypertarget{method-OutputParser-clean}{}}}
\subsection{Method \code{clean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputParser$clean(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputParser-content_detect"></a>}}
\if{latex}{\out{\hypertarget{method-OutputParser-content_detect}{}}}
\subsection{Method \code{content_detect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputParser$content_detect(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector.
Parse raw TNT tree output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputParser-transform"></a>}}
\if{latex}{\out{\hypertarget{method-OutputParser-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputParser$transform(output, content_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{A character vector of raw TNT output.}

\item{\code{content_type}}{A character vector indicating the content type of \code{output}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputParser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OutputParser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputParser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
