% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCharacterMatrix.R
\name{print.MultiCharacterMatrix}
\alias{print.MultiCharacterMatrix}
\title{Print MultiCharacterMatrix object}
\usage{
\method{print}{MultiCharacterMatrix}(x, ...)
}
\arguments{
\item{x}{A MultiCharacterMatrix object.}

\item{...}{Ignored.}
}
\description{
Print MultiCharacterMatrix object
}
