% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonophyleticConstraint-class.R
\name{print.MultiMonophyleticConstraint}
\alias{print.MultiMonophyleticConstraint}
\title{Print a MultiMonophyleticConstraint Object}
\usage{
\method{print}{MultiMonophyleticConstraint}(x, ...)
}
\arguments{
\item{x}{A \code{MonophyleticConstraint} object.}

\item{...}{Ignored.}
}
\description{
Print a MultiMonophyleticConstraint Object
}
\keyword{internal}
