% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonophylyConstraint-class.R
\name{print.MultiMonophylyConstraint}
\alias{print.MultiMonophylyConstraint}
\title{Print a MultiMonophylyConstraint Object}
\usage{
\method{print}{MultiMonophylyConstraint}(x, ...)
}
\arguments{
\item{x}{A \code{MonophylyConstraint} object.}

\item{...}{Ignored.}
}
\description{
Print a MultiMonophylyConstraint Object
}
\keyword{internal}
