% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueueGenerator-class.R
\name{QueueGenerator}
\alias{QueueGenerator}
\title{Queue generator}
\description{
Defines functions common to objects that generate queues.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QueueGenerator-queue}{\code{QueueGenerator$queue()}}
\item \href{#method-QueueGenerator-clone}{\code{QueueGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueueGenerator-queue"></a>}}
\if{latex}{\out{\hypertarget{method-QueueGenerator-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueGenerator$queue(.queue = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.queue}}{A \code{\link{CommandQueue}} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueueGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QueueGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
