% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomSectorialSearchOptions-class.R
\name{RandomSectorialSearchOptions}
\alias{RandomSectorialSearchOptions}
\title{Set options for a random sectorial search}
\description{
\code{RandomSectorialSearchOptions} is an R6 class that defines the set of
  options required to perform sectorial searching analyses in \code{nitro}.
}
\section{Super class}{
\code{\link[nitro:SectorialSearchBaseOptions]{nitro::SectorialSearchBaseOptions}} -> \code{RandomSectorialSearchOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{min_size}}{An integer value indicating the minimum size of random
selections. If \code{NULL} (the default), this value will be
automatically set according to the number of taxa in the matrix when
used with a \code{"\link{TreeAnalysis}"} object.}

\item{\code{max_size}}{An integer value indicating the maximum size of random
selections. If \code{NULL} (the default), this value will be
automatically set according to the number of taxa in the matrix when a
new \code{"\link{TreeAnalysis}"} object is initialized.}

\item{\code{selection_factor}}{An integer value indicating the maximum number
of random selections for the currently active taxa.}

\item{\code{increase}}{An integer value indicating the factor to increase the
size of random selections if enough selection of the current size have
been completed.}

\item{\code{small_starts}}{An integer value indicating the number of random
addition sequence plus tree-bisection reconnection swaps to perform
for random selections below \code{min_size}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RandomSectorialSearchOptions-new}{\code{RandomSectorialSearchOptions$new()}}
\item \href{#method-RandomSectorialSearchOptions-print}{\code{RandomSectorialSearchOptions$print()}}
\item \href{#method-RandomSectorialSearchOptions-queue}{\code{RandomSectorialSearchOptions$queue()}}
\item \href{#method-RandomSectorialSearchOptions-clone}{\code{RandomSectorialSearchOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomSectorialSearchOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-RandomSectorialSearchOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomSectorialSearchOptions$new(
  min_size = NULL,
  max_size = NULL,
  selection_factor = 50,
  increase = 100,
  small_starts = 3,
  buffer = TRUE,
  slack = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_size}}{An integer value indicating the minimum size of random
selections. If left as 0 (the default), this value will be
automatically set according to the number of taxa in the matrix when a
new \code{"\link{TreeAnalysis}"} object is initialized.}

\item{\code{max_size}}{An integer value indicating the maximum size of random
selections. If left at 0 (the default), this value will be
automatically set according to the number of taxa in the matrix when a
new \code{"\link{TreeAnalysis}"} object is initialized.}

\item{\code{selection_factor}}{An integer value indicating the maximum number
of random selections for the currently active taxa.}

\item{\code{increase}}{An integer value indicating the factor to increase the
size of random selections if enough selection of the current size have
been completed.}

\item{\code{small_starts}}{An integer value indicating the number of random
addition sequence plus tree-bisection reconnection swaps to perform
for random selections below \code{min_size}.}

\item{\code{buffer}}{A logical value indicating whether to use an independent}

\item{\code{slack}}{An integer value indicating the percentage to increase the
available memory during searches.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomSectorialSearchOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-RandomSectorialSearchOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomSectorialSearchOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomSectorialSearchOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-RandomSectorialSearchOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomSectorialSearchOptions$queue(set_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_only}}{A logical indicating whether to instruct the command to
execute immediately (\code{TRUE}) or set the variables for future
execution \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomSectorialSearchOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RandomSectorialSearchOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomSectorialSearchOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
