% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatchetOptions-class.R
\name{RatchetOptions}
\alias{RatchetOptions}
\title{Set options for parsimony ratchet}
\description{
\code{RatchetOptions} is an R6 class that defines the set of options
for performing a parsimony ratchet phylogenetic analysis in \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iterations}}{An integer value indicating the number of iterations.}

\item{\code{replacements}}{An integer value indicating the number of
replacements (i.e., accepted tree rearrangements) to perform in each
perturbation phase.}

\item{\code{prob_up}}{An integer value indicating the probability of
upweighting a character.}

\item{\code{prob_down}}{An integer value indicating the probability of
downweighting a character.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RatchetOptions-new}{\code{RatchetOptions$new()}}
\item \href{#method-RatchetOptions-print}{\code{RatchetOptions$print()}}
\item \href{#method-RatchetOptions-queue}{\code{RatchetOptions$queue()}}
\item \href{#method-RatchetOptions-clone}{\code{RatchetOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatchetOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-RatchetOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatchetOptions$new(
  iterations = 50,
  replacements = 40,
  prob_up = 4,
  prob_down = 4
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{An integer value indicating the number of iterations.}

\item{\code{replacements}}{An integer value indicating the number of
replacements (i.e., accepted tree rearrangements) to perform in each
perturbation phase.}

\item{\code{prob_up}}{An integer value indicating the probability of
upweighting a character.}

\item{\code{prob_down}}{An integer value indicating the probability of
downweighting a character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatchetOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-RatchetOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatchetOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatchetOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-RatchetOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatchetOptions$queue(set_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_only}}{A logical indicating whether to instruct the command to
execute immediately (\code{TRUE}) or set the variables for future
execution (\code{FALSE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RatchetOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RatchetOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatchetOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
