% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchSwap-methods.R,
%   R/NitroDriven-methods.R, R/NitroResampleBase-methods.R
\name{replications}
\alias{replications}
\alias{replications,NitroBranchSwap-method}
\alias{replications<-}
\alias{replications<-,NitroBranchSwap,numeric-method}
\alias{replications,NitroDriven-method}
\alias{replications<-,NitroDriven,numeric-method}
\alias{replications,NitroResampleBase-method}
\title{Replications}
\usage{
replications(n)

\S4method{replications}{NitroBranchSwap}(n)

replications(n) <- value

\S4method{replications}{NitroBranchSwap,numeric}(n) <- value

\S4method{replications}{NitroDriven}(n)

\S4method{replications}{NitroDriven,numeric}(n) <- value

\S4method{replications}{NitroResampleBase}(n)
}
\arguments{
\item{n}{an object of either \code{NitroBranchSwap} or \code{NitroDriven}.}

\item{value}{a numeric vector indicating the number of replications for the
analysis.}
}
\value{
a numeric vector indicating the number of replications for the
analysis.
}
\description{
Function to return or set the number of replications for a branch swapping
or driven phylogenetic analysis
}
