% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroResampleBase-class.R
\name{ResampleBaseOptions}
\alias{ResampleBaseOptions}
\title{Set options common to resampling analyses}
\description{
\code{ResampleBaseOptions} is an R6 class that defines options common to
  resampling analyses.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{search_method}}{A valid tree search configuration.}

\item{\code{replications}}{An integer value indicating the number of resampling
replications to perform.}

\item{\code{cutoff}}{An integer value indicating the cutoff value for
frequencies.}

\item{\code{frequency_summary}}{A character vector indicating which method(s) to
  use to summarize supports. More than one option can be specified. The
  options are:
\itemize{
  \item \code{difference}: frequency differences (i.e., group supported/contradicted), default;
  \item \code{absolute}: absolute frequencies; or
  \item \code{slope}: frequency slopes.
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResampleBaseOptions-new}{\code{ResampleBaseOptions$new()}}
\item \href{#method-ResampleBaseOptions-clone}{\code{ResampleBaseOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResampleBaseOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResampleBaseOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResampleBaseOptions$new(
  search_method,
  replications = 100,
  frequency_summary = "absolute"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_method}}{A valid tree search configuration.}

\item{\code{replications}}{An integer value indicating the number of resampling
replications to perform.}

\item{\code{frequency_summary}}{A character vector indicating which method(s) to
  use to summarize supports. More than one option can be specified. The
  options are:
\itemize{
  \item \code{difference}: frequency differences (i.e., group supported/contradicted), detault;
  \item \code{absolute}: absolute frequencies; or
  \item \code{slope}: frequency slopes.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResampleBaseOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResampleBaseOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResampleBaseOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
