% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_command.R
\name{resolve_command}
\alias{resolve_command}
\title{Resolve a single command}
\usage{
resolve_command(command, resolved, type = "required", suppress_errors = TRUE)
}
\arguments{
\item{command}{A command to be resolved}

\item{resolved}{A \code{CommandList} of resolved commands}

\item{type}{The type of dependency to resolve}

\item{suppress_errors}{[\code{logical(1)}]\cr
If \code{TRUE} (default), unmet dependencies return \code{FALSE} silently.
If \code{FALSE}, an error is raised immediately.}
}
\description{
Resolve a single command
}
\keyword{internal}
