% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTnt.R
\name{runTnt}
\alias{runTnt}
\title{Executes TNT commands}
\usage{
runTnt(tnt.path, matrix, cmds)
}
\arguments{
\item{tnt.path}{The location of the TNT command line binary.}

\item{matrix}{A \code{phyDat} object of the matrix.}

\item{cmds}{A vector of commands to be issued by TNT.}
}
\value{
A character vector of the output from TNT.
}
\description{
Executes commands on a supplied phylogenetic matrix in the TNT command line binary.
}
