% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srchmd.R
\name{schmd}
\alias{schmd}
\title{Syntactic Sugar for Search Method Construction}
\usage{
schmd(name, ...)
}
\arguments{
\item{name}{The name of a search method. Possible values are:
\itemize{
  \item \code{\link[nitro:NitroBootstrap]{Bootstrap}}
  \item \code{\link[nitro:NitroBranchBreak]{BranchSwap}}
  \item \code{\link[nitro:NitroBranchSwap]{BranchSwap}}
  \item \code{\link[nitro:NitroConstraintSectorialSearch]{ConstraintSectorialSearch}}
  \item \code{\link[nitro:NitroDriven]{Driven}}
  \item \code{\link[nitro:NitroImplicitEnum]{ImplicitEnum}}
  \item \code{\link[nitro:NitroJackknife]{Jackknife}}
  \item \code{\link[nitro:NitroRandomSectorialSearch]{RandomSectorialSearch}}
  \item \code{\link[nitro:NitroRatchet]{Ratchet}}
  \item \code{\link[nitro:NitroSymmetricResample]{SymmetricResample}}
  \item \code{\link[nitro:NitroTreeDrift]{TreeDrift}}
  \item \code{\link[nitro:NitroTreeFuse]{TreeFuse}}
  \item \code{\link[nitro:NitroTreeHybridize]{TreeHybridize}}
}}

\item{...}{Named arguments passed to the constructor for the search method,
to be set as parameters.}
}
\description{
A shortcut function for retrieving tree search objects (i.e., those
inheriting from \code{"\link{NitroMethodsBase}"} and search parameter
definition.
}
