% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenSizeCommand-class.R
\name{ScreenSizeCommand}
\alias{ScreenSizeCommand}
\title{Screen Size Command}
\description{
An \link[R6:R6Class]{R6} class that configures the dimensions of the TNT
screen buffer.

This command is created automatically as part of the standard
\link{CommandQueue} setup. Users do not typically need to instantiate it
directly.
}
\details{
\subsection{Command output}{

\verb{$render()} produces a string of the form \verb{screen \{rows\}x\{columns\};}
}

\subsection{Queue integration}{

Calling \verb{$enqueue()} adds this command to a \link{CommandQueue} at priority
\code{110}.
}
}
\keyword{internal}
\section{Super class}{
\code{\link[nitro:BasicCommand]{BasicCommand}} -> \code{ScreenSizeCommand}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{columns}}{[\code{integer(1)}]\cr
The number of columns in the screen buffer. Must be a positive
integer.}

    \item{\code{rows}}{[\code{integer(1)}]\cr
The number of rows in the screen buffer. Must be a positive
integer.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ScreenSizeCommand-enqueue}{\code{ScreenSizeCommand$enqueue()}}
    \item \href{#method-ScreenSizeCommand-initialize}{\code{ScreenSizeCommand$new()}}
    \item \href{#method-ScreenSizeCommand-render}{\code{ScreenSizeCommand$render()}}
    \item \href{#method-ScreenSizeCommand-clone}{\code{ScreenSizeCommand$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="format"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-format'><code>BasicCommand$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="get_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-get_dependency'><code>BasicCommand$get_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="new_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-new_dependency'><code>BasicCommand$new_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="print"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-print'><code>BasicCommand$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="set_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-set_dependency'><code>BasicCommand$set_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="transform"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-transform'><code>BasicCommand$transform()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScreenSizeCommand-enqueue"></a>}}
\if{latex}{\out{\hypertarget{method-ScreenSizeCommand-enqueue}{}}}
\subsection{\code{ScreenSizeCommand$enqueue()}}{
  Add this command to a \link{CommandQueue}.

Adds this command at priority \code{110} by default.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScreenSizeCommand$enqueue(.queue, priority = 110)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{.queue}}{A \link{CommandQueue} object.}
      \item{\code{priority}}{[\code{integer(1)}]\cr
A non-negative integer controlling execution order. Lower values
execute first.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScreenSizeCommand-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ScreenSizeCommand-initialize}{}}}
\subsection{\code{ScreenSizeCommand$new()}}{
  Create a new \code{ScreenSizeCommand} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScreenSizeCommand$new(columns, rows)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{columns}}{[\code{integer(1)}]\cr
The number of columns in the screen buffer. See the \verb{$columns}
field.}
      \item{\code{rows}}{[\code{integer(1)}]\cr
The number of rows in the screen buffer. See the \verb{$rows} field.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{ScreenSizeCommand} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScreenSizeCommand-render"></a>}}
\if{latex}{\out{\hypertarget{method-ScreenSizeCommand-render}{}}}
\subsection{\code{ScreenSizeCommand$render()}}{
  Render the TNT \code{screen} command string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScreenSizeCommand$render(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Not used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A single-element character vector containing the TNT
command.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScreenSizeCommand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScreenSizeCommand-clone}{}}}
\subsection{\code{ScreenSizeCommand$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ScreenSizeCommand$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
