% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{search_method}
\alias{search_method}
\title{Return search method}
\usage{
search_method(n)
}
\arguments{
\item{n}{a \code{\link{NitroTreeSearch}} object.}
}
\description{
A function that returns and sets the current search method
}
