% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{start_trees}
\alias{start_trees}
\alias{start_trees,NitroTreeSearch-method}
\alias{start_trees<-}
\alias{start_trees<-,NitroTreeSearch-method}
\title{Starting trees}
\usage{
start_trees(n)

\S4method{start_trees}{NitroTreeSearch}(n)

start_trees(n) <- value

\S4method{start_trees}{NitroTreeSearch}(n) <- value
}
\arguments{
\item{n}{a \code{"\linkS4class{NitroTreeSearch}"} object}

\item{value}{a \code{multiPhylo} list.}
}
\value{
a \code{multiPhylo} list.
}
\description{
A function that returns or sets the starting trees.
}
