% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroSymmetricResample-class.R
\name{SymmetricResamplingOptions}
\alias{SymmetricResamplingOptions}
\title{Set options for symmetric resampling}
\description{
\code{SymmetricResamplingOptions} is an R6 class that defines parameters for
symmetric resampling analyses.
}
\section{Super class}{
\code{\link[nitro:ResampleBaseOptions]{nitro::ResampleBaseOptions}} -> \code{SymmetricResamplingOptions}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{probability}}{An integer value indicating the change probability.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SymmetricResamplingOptions-new}{\code{SymmetricResamplingOptions$new()}}
\item \href{#method-SymmetricResamplingOptions-print}{\code{SymmetricResamplingOptions$print()}}
\item \href{#method-SymmetricResamplingOptions-queue}{\code{SymmetricResamplingOptions$queue()}}
\item \href{#method-SymmetricResamplingOptions-clone}{\code{SymmetricResamplingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymmetricResamplingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-SymmetricResamplingOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymmetricResamplingOptions$new(
  probability = 36,
  search_method = NULL,
  replications = 100,
  cutoff = 0,
  frequency_summary = "absolute"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probability}}{An integer value indicating the change probability.}

\item{\code{search_method}}{An object inheriting class
\code{"\link{NitroMethodsBase}"}.}

\item{\code{replications}}{An integer value indicating the number of resampling
replications to perform.}

\item{\code{cutoff}}{An integer value indicating the cutoff value for
frequencies.}

\item{\code{frequency_summary}}{A character vector indicating which method(s) to
  use to summarize supports. More than one option can be specified. The
  options are:
\itemize{
  \item \code{absolute}: absolute frequencies, default;
  \item \code{difference}: frequency differences (i.e., group supported/contradicted);
  \item \code{slope}: frequency slopes.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymmetricResamplingOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-SymmetricResamplingOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymmetricResamplingOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymmetricResamplingOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-SymmetricResamplingOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymmetricResamplingOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymmetricResamplingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SymmetricResamplingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymmetricResamplingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
