% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBase-methods.R
\name{inactive_taxa}
\alias{inactive_taxa}
\alias{inactive_taxa,NitroBase,missing-method}
\alias{inactive_taxa<-}
\alias{inactive_taxa<-,NitroBase,character-method}
\alias{inactive_taxa<-,NitroBase,NULL-method}
\title{List inactive taxa}
\usage{
inactive_taxa(n, value)

\S4method{inactive_taxa}{NitroBase,missing}(n)

inactive_taxa(n) <- value

\S4method{inactive_taxa}{NitroBase,character}(n) <- value

\S4method{inactive_taxa}{NitroBase,`NULL`}(n) <- value
}
\arguments{
\item{n}{an object that inherits \code{NitroBase}.}

\item{value}{a character vector indicating the taxa to be marked as inactive.}
}
\value{
a character vector indicating the taxa defined as inactive.
}
\description{
Function that returns the names of taxa defined as inactive.
}
