% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTnt.R
\name{tnt}
\alias{tnt}
\title{Executes TNT commands}
\usage{
tnt(params, tnt.path)
}
\arguments{
\item{params}{the parameters for an analysis as specified from the output
of the \code{implicit.enum}, \code{branchswap}, \code{ratchet}, or
\code{driven} commands.}

\item{tnt.path}{the location of the TNT command line binary.}
}
\value{
An object containing the TNT parameters and the \code{phyDat} matrix
  used in the analysis, and a \code{multiPhylo} containing the trees found
  during the search.
}
\description{
Executes commands on a supplied phylogenetic matrix in the TNT
  command line binary.
}
