% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt_attach.R
\name{tnt_attach}
\alias{tnt_attach}
\title{Register TNT executable}
\usage{
tnt_attach(path, .envir = parent.frame())
}
\arguments{
\item{path}{The location of the TNT command line executable.}

\item{.envir}{The environment that the TNTProcess object will be assigned to.}
}
\description{
\code{tnt_attach} attaches an instance of TNT to an R environment.
}
