% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBase-methods.R,
%   R/NitroBranchSwap-methods.R, R/NitroDriven-methods.R,
%   R/NitroImplicitEnum-methods.R, R/NitroRatchet-methods.R
\name{tnt_cmd}
\alias{tnt_cmd}
\alias{tnt_cmd,NitroBranchSwap-method}
\alias{tnt_cmd,NitroDriven-method}
\alias{tnt_cmd,NitroImplicitEnum-method}
\alias{tnt_cmd,NitroRatchet-method}
\title{Return TNT command}
\usage{
tnt_cmd(n)

\S4method{tnt_cmd}{NitroBranchSwap}(n)

\S4method{tnt_cmd}{NitroDriven}(n)

\S4method{tnt_cmd}{NitroImplicitEnum}(n)

\S4method{tnt_cmd}{NitroRatchet}(n)
}
\arguments{
\item{n}{an object that inherits from class \code{NitroBase}.}
}
\value{
a character vector of the TNT command.
}
\description{
A function that returns the command to perform the phylogenetic analysis for
a branch swapping analysis.
}
