% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt-output-parse.R
\name{tnt_output_parse}
\alias{tnt_output_parse}
\title{Extracts trees from TNT output.}
\usage{
tnt_output_parse(tnt_output, tip_labels)
}
\arguments{
\item{tnt_output}{a character vector containing raw TNT output from the
\code{implicit.enum}, \code{branchswap}, \code{ratchet} or \code{driven}
functions.}

\item{tip_labels}{a list or vector containing the names of the taxa from the
analysed matrix.}
}
\value{
a \code{multiPhylo} object.
}
\description{
Extracts trees from TNT output.
}
