% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNTProcess-class.R
\name{TNTProcess}
\alias{TNTProcess}
\title{TNT executable interface object}
\description{
\code{TNTProcess} is an R6 class that starts and manages an instance of TNT.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{platform}}{The platform the TNT executable is running on.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TNTProcess-execute_queue}{\code{TNTProcess$execute_queue()}}
\item \href{#method-TNTProcess-new}{\code{TNTProcess$new()}}
\item \href{#method-TNTProcess-is_active}{\code{TNTProcess$is_active()}}
\item \href{#method-TNTProcess-print}{\code{TNTProcess$print()}}
\item \href{#method-TNTProcess-clone}{\code{TNTProcess$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TNTProcess-execute_queue"></a>}}
\if{latex}{\out{\hypertarget{method-TNTProcess-execute_queue}{}}}
\subsection{Method \code{execute_queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TNTProcess$execute_queue(queue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queue}}{A \link{"\code{CommandQueue}"} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TNTProcess-new"></a>}}
\if{latex}{\out{\hypertarget{method-TNTProcess-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TNTProcess$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The location of the TNT command line executable.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TNTProcess-is_active"></a>}}
\if{latex}{\out{\hypertarget{method-TNTProcess-is_active}{}}}
\subsection{Method \code{is_active()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TNTProcess$is_active(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TNTProcess-print"></a>}}
\if{latex}{\out{\hypertarget{method-TNTProcess-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TNTProcess$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TNTProcess-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TNTProcess-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TNTProcess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
