% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntTreeParse.R
\name{tntTreeParse}
\alias{tntTreeParse}
\title{Extracts trees from TNT output.}
\usage{
tntTreeParse(tnt.output, tip.labels)
}
\arguments{
\item{tnt.output}{a character vector containing raw TNT output from the
\code{implicit.enum}, \code{branchswap}, \code{ratchet} or \code{driven}
functions.}

\item{tip.labels}{a list or vector containing the names of the taxa from the
analysed matrix.}
}
\value{
a \code{multiPhylo} object.
}
\description{
Extracts trees from TNT output.
}
