% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFusingOptions-class.R
\name{TreeFusingOptions}
\alias{TreeFusingOptions}
\title{Set options for tree fusing}
\description{
\code{TreeFusingOptions} is an R6 class that defines the set of parameters
required for performing tree fusing operations in \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of tree-fusing rounds
to perform.}

\item{\code{exchange_equal}}{A logical value indicating whether to accept exchanges
of equal score.}

\item{\code{start_best}}{A logical value indicating whether to use the best tree to
start tree-fusing.}

\item{\code{keep_all}}{A logical value indicating whether to keep all trees found
instead of only the best trees.}

\item{\code{accept_all}}{A logical value indicating whether to accept all exchanges
rather than only those that improve the best score.}

\item{\code{swap}}{A logical value indicating whether to perform tree-bisection
reconnection swapping after exchanging clades.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreeFusingOptions-new}{\code{TreeFusingOptions$new()}}
\item \href{#method-TreeFusingOptions-print}{\code{TreeFusingOptions$print()}}
\item \href{#method-TreeFusingOptions-queue}{\code{TreeFusingOptions$queue()}}
\item \href{#method-TreeFusingOptions-clone}{\code{TreeFusingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeFusingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreeFusingOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeFusingOptions$new(
  rounds = 5,
  exchange_equal = FALSE,
  start_best = TRUE,
  keep_all = TRUE,
  accept_all = TRUE,
  swap = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of tree-fusing rounds
to perform.}

\item{\code{exchange_equal}}{A logical value indicating whether to accept exchanges
of equal score.}

\item{\code{start_best}}{A logical value indicating whether to use the best tree to
start tree-fusing.}

\item{\code{keep_all}}{A logical value indicating whether to keep all trees found
instead of only the best trees.}

\item{\code{accept_all}}{A logical value indicating whether to accept all exchanges
rather than only those that improve the best score.}

\item{\code{swap}}{A logical value indicating whether to perform tree-bisection
reconnection swapping after exchanging clades.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeFusingOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-TreeFusingOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeFusingOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeFusingOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-TreeFusingOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeFusingOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeFusingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreeFusingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeFusingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
