% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeHybridizingOptions-class.R
\name{TreeHybridizingOptions}
\alias{TreeHybridizingOptions}
\title{Set options for tree hybridizing}
\description{
\code{TreeHybridizingOptions} is an R6 class that defines the set of
  parameters required for performing tree hybridizing operations in
  \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of rounds of
tree-hybridizing to perform.}

\item{\code{hybridizations}}{An integer value indicating the number of
hybridizations to perform in each round.}

\item{\code{best_trees}}{An integer value indicating the number of best trees from
the previous round of hybridizing to use in the next round.}

\item{\code{replace}}{A logical value indicating whether to replace the source tree
with a better tree produced by hybridizing.}

\item{\code{sample_factor}}{An integer value indicating the number of times to
increase the size of initial tree set by. The corresponding number of trees
to retain will be proportional to the inverse of this value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreeHybridizingOptions-new}{\code{TreeHybridizingOptions$new()}}
\item \href{#method-TreeHybridizingOptions-print}{\code{TreeHybridizingOptions$print()}}
\item \href{#method-TreeHybridizingOptions-queue}{\code{TreeHybridizingOptions$queue()}}
\item \href{#method-TreeHybridizingOptions-clone}{\code{TreeHybridizingOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeHybridizingOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreeHybridizingOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeHybridizingOptions$new(
  rounds = 1,
  hybridizations = 1000,
  best_trees = 50,
  replace = TRUE,
  sample_factor = 15
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of rounds of
tree-hybridizing to perform.}

\item{\code{hybridizations}}{An integer value indicating the number of
hybridizations to perform in each round.}

\item{\code{best_trees}}{An integer value indicating the number of best trees from
the previous round of hybridizing to use in the next round.}

\item{\code{replace}}{A logical value indicating whether to replace the source tree
with a better tree produced by hybridizing.}

\item{\code{sample_factor}}{An integer value indicating the number of times to
increase the size of initial tree set by. The corresponding number of trees
to retain will be proportional to the inverse of this value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeHybridizingOptions-print"></a>}}
\if{latex}{\out{\hypertarget{method-TreeHybridizingOptions-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeHybridizingOptions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeHybridizingOptions-queue"></a>}}
\if{latex}{\out{\hypertarget{method-TreeHybridizingOptions-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeHybridizingOptions$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeHybridizingOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreeHybridizingOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeHybridizingOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
