% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTrees-methods.R
\name{trees}
\alias{trees}
\alias{trees,NitroTrees-method}
\title{Get trees}
\usage{
trees(n)

\S4method{trees}{NitroTrees}(n)
}
\arguments{
\item{n}{a \code{NitroTrees} object}
}
\description{
A function to return the trees from a \code{"\linkS4class{NitroTrees}"}
object.
}
