% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTNTNexus.R
\name{writeTNTNexus}
\alias{writeTNTNexus}
\title{Write a TNT-compatible nexus file}
\usage{
writeTNTNexus(obj, filename, read_trees = FALSE)
}
\arguments{
\item{obj}{an object that inherits \code{"\linkS4class{NitroTreeSearch}"}.}

\item{filename}{a character vector indicating the name for the new file.}

\item{read_trees}{a logical value indicating whether to read in trees from
the \code{start_trees} slot of \code{obj} prior to analysis.}
}
\description{
Function to write a nexus file from a \code{"\linkS4class{NitroTreeSearch}"}
object, containing a phylogenetic matrix and TNT tree search commands
required as required for an analysis in \code{nitro}. This is used
internally by \code{\link{tnt}}.
}
\details{
The function goes through the following stages, in order:

\enumerate{
\item A temporary TNT-compatible nexus file is created from the input
  matrix.
\item A TNT block is appended to this file containing a script that will
  execute in order each step required to perform the phylogenetic analysis
  and return the set of most parsimonious trees generated.

  The script generation step itself involves the following steps (not all of
  which will necessarily be present depending on the specifications of the
  given analysis):

  \enumerate{
  \item set implied weighting constants and parameters;
  \item set zero-length branch collapse rule;
  \item set maximum number of trees to hold in TNT's buffer;
  \item set the outgroup taxon;
  \item defining ordered and inactive characters;
  \item defining inactive taxa;
  \item defining and setting constraints on monophyly;
  \item reading in starting trees;
  \item setting parameters for and executing tree searches;
  \item condensing trees according to current collapse rule;
  \item retrieving trees as character strings; and
  \item retrieving tree lengths/scores and minimum-maximum scores.
  }
\item This file will then be executed by a subprocess of TNT created using
  \code{processx}, and the output generated from the \code{stdout} stream
  of this process is captured.
\item Tree strings and metrics are parsed from the output stream and
  converted into a \code{multiPhylo} object. During this step, the ensemble
  consistency, retention and rescaled consistency indices (CI, RI, RC) are
  calculated from the output. These values, together with the tree
  lengths/scores, are saved with each \code{phylo} tree.
\item A \code{"\linkS4class{NitroTrees}"} object is returned which contains
  a \code{multiPhylo} object in the \code{trees} slot and the original
  \code{"\linkS4class{NitroTreeSearch}"} object in the \code{tree_search}
  slot.
}
}
