% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCharacterMatrix-class.R
\name{AbstractCharacterMatrix}
\alias{AbstractCharacterMatrix}
\title{Abstract character matrix}
\description{
\code{AbstractCharacterMatrix} is an R6 class that defines properties common
  to any type of character matrix. Intended to be inherited, not instantiated
  directly.
}
\section{Super classes}{
\code{\link[nitro:QueueGenerator]{nitro::QueueGenerator}} -> \code{\link[nitro:AbstractModule]{nitro::AbstractModule}} -> \code{AbstractCharacterMatrix}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{The character matrix.}

\item{\code{taxa}}{The names of the taxa contained in the matrix.}

\item{\code{n_characters}}{The number of the characters contained in the matrix.}

\item{\code{inactive}}{A numeric vector indicating which characters to mark as inactive.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractCharacterMatrix-clone}{\code{AbstractCharacterMatrix$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nitro" data-topic="QueueGenerator" data-id="queue"><a href='../../nitro/html/QueueGenerator.html#method-QueueGenerator-queue'><code>nitro::QueueGenerator$queue()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractCharacterMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractCharacterMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractCharacterMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
