% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintsBase-methods.R
\name{addConstraint}
\alias{addConstraint}
\title{Add new constraint}
\usage{
addConstraint(
  tree_search,
  fixed_taxa,
  floating_taxa = character(),
  type = c("positive", "negative")
)
}
\arguments{
\item{fixed_taxa}{a character vector of taxa to set as fixed constraints}

\item{floating_taxa}{a character vector of taxa to set as floating
constraints}

\item{group}{an object of type \code{\link{NitroConstraint}}}
}
\description{
Function to add a new constraint to a constraint group.
}
