% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchswap.R
\name{branchswap}
\alias{branchswap}
\title{Phylogenetic analysis using branch swapping}
\usage{
branchswap(
  matrix,
  collapse = 3,
  hold = 100,
  outgroup = NULL,
  replications = 10,
  hold.rep = 10,
  keepall = FALSE
)
}
\arguments{
\item{matrix}{a \code{phyDat} object of the matrix.}

\item{collapse}{an integer indicating the rule for collapsing of zero length
branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}

\item{hold}{an integer value indicating the maximum number of trees to allow
TNT to hold in memory.}

\item{outgroup}{the name of the taxon to set as the outgroup for the
phylogenetic analysis. By default, the first taxon in the matrix is the
outgroup.}

\item{replications}{an integer value indicating the number of replications.}

\item{hold.rep}{an integer value inficating the maximum number of trees to
retain during each replication.}

\item{keepall}{a logical value indicating whether to retain all generated
trees from each replication regardless of length.}
}
\value{
a list containing the search parameters and TNT command string, a
  \code{phyDat} object containing the phylogenetic matrix analysed and, if
  \code{run.now} is \code{TRUE}, a \code{multiPhylo} object of trees found
  from the search commands.
}
\description{
Phylogenetic analysis using branch swapping
}
