% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canale_2022}
\alias{canale_2022}
\title{Canale et al. (2022) matrix}
\format{
A phyDat object containing 20 taxa and 175 characters.
}
\source{
Downloaded from the Supplemental Information section of
  Canale et al. (2022) <https://doi.org/10.1016/j.cub.2022.05.057>,
  converted into a Nexus file in Mesquite 3.61, and read with ReadAsPhydat
  from TreeTools.
}
\usage{
canale_2022
}
\description{
A phylogenetic matrix of carcharodontosaurid theropod dinosaurs.
}
\keyword{datasets}
