% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBase-methods.R
\name{ordered_characters}
\alias{ordered_characters}
\alias{ordered_characters,NitroBase-method}
\alias{ordered_characters<-}
\alias{ordered_characters<-,NitroBase,numeric-method}
\alias{ordered_characters<-,NitroBase,NULL-method}
\alias{inactive_characters}
\alias{inactive_characters,NitroBase-method}
\alias{inactive_characters<-}
\alias{inactive_characters<-,NitroBase,numeric-method}
\alias{inactive_characters<-,NitroBase,NULL-method}
\title{Character modification functions}
\usage{
ordered_characters(n)

\S4method{ordered_characters}{NitroBase}(n)

ordered_characters(n) <- value

\S4method{ordered_characters}{NitroBase,numeric}(n) <- value

\S4method{ordered_characters}{NitroBase,`NULL`}(n) <- value

inactive_characters(n)

\S4method{inactive_characters}{NitroBase}(n)

inactive_characters(n) <- value

\S4method{inactive_characters}{NitroBase,numeric}(n) <- value

\S4method{inactive_characters}{NitroBase,`NULL`}(n) <- value
}
\arguments{
\item{n}{an object that inherits \code{NitroBase}.}

\item{value}{a numeric vector indicating the characters to be marked as inactive.}
}
\value{
a numeric vector indicating the characters defined as ordered.

a numeric vector indicating the characters defined as inactive.
}
\description{
List ordered characters

List inactive characters
}
\details{
Function to return or set indexes of characters as ordered.

Function that returns the indexes of characters defined as inactive.
}
