% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousMatrix-class.R
\name{ContinuousMatrix}
\alias{ContinuousMatrix}
\title{Create a discrete character matrix}
\description{
\code{ContinuousMatrix} is an R6 class that contains a continuous character
  matrix and functions for modifying character activity.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{taxa}}{The names of the taxa contained in the matrix.}

\item{\code{n_characters}}{The number of the characters contained in the matrix.}

\item{\code{inactive}}{A numeric vector indicating which characters to mark as inactive.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContinuousMatrix-new}{\code{ContinuousMatrix$new()}}
\item \href{#method-ContinuousMatrix-print}{\code{ContinuousMatrix$print()}}
\item \href{#method-ContinuousMatrix-queue}{\code{ContinuousMatrix$queue()}}
\item \href{#method-ContinuousMatrix-clone}{\code{ContinuousMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousMatrix-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousMatrix$new(matrix, inactive = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{An \code{phyDat} discrete character matrix.}

\item{\code{inactive}}{A numeric vector indicating which characters to mark as inactive.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousMatrix-print"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousMatrix-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousMatrix$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousMatrix-queue"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousMatrix-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousMatrix$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
