% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driven.R
\name{driven}
\alias{driven}
\title{Phylogenetic analysis using driven (or "New Technology") search}
\usage{
driven(
  matrix,
  collapse = 3,
  hold = 100,
  outgroup = NULL,
  hits = 1,
  replications = 4,
  hold.rep = 1,
  rss = TRUE,
  css = FALSE,
  xss = FALSE,
  ratchet.cycles = 0,
  drifting.cycles = 30,
  fusing.rounds = 0,
  consense.times = 0,
  multiply = FALSE,
  keepall = FALSE
)
}
\arguments{
\item{matrix}{a \code{phyDat} object of the matrix.}

\item{collapse}{an integer indicating the rule for collapsing of zero length
branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}

\item{hold}{an integer value indicating the maximum number of trees to allow
TNT to hold in memory.}

\item{outgroup}{the name of the taxon to set as the outgroup for the
phylogenetic analysis. By default, the first taxon in the matrix is the
outgroup.}

\item{hits}{an integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{replications}{an integer value indicating the number of replications.}

\item{hold.rep}{an integer value inficating the maximum number of trees to
retain during each replication.}

\item{rss, css, xss}{a logival value indicating whether to use random,
constraint or exclusive sectorial searches.}

\item{ratchet.cycles}{an integer value indicating the number of cycles of
the parsimony ratchet to perform per replication.}

\item{drifting.cycles}{an integer value indicating the number of cycles of
tree drifting to perform per replication.}

\item{fusing.rounds}{an integer value indicating the number of rounds of
fusing per replication.}

\item{consense.times}{an integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{multiply}{a logical value indicating whether to find additional trees
by fusing suboptimal with optimal trees after hitting the target score.}

\item{keepall}{a logical value indicating whether to retain all generated
trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When
\code{hits} = 1, it is trees from each of the RAS + TBR +  SS or DFT or
RAT, in addition to the trees resulting from fusing those. When
\code{hits} > 1, then it means the trees resulting from fusing the
initial starting trees for each of starting points.}
}
\value{
a list containing the search parameters and TNT command string, a
  \code{phyDat} object containing the phylogenetic matrix analysed and, if
  \code{run.now} is \code{TRUE}, a \code{multiPhylo} object of trees found
  from the search commands.
}
\description{
Phylogenetic analysis using driven (or "New Technology") search
}
