% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{inactiveCharacters}
\alias{inactiveCharacters}
\title{Define inactive characters}
\usage{
inactiveCharacters(matrix, characters)
}
\arguments{
\item{matrix}{A \code{phyDat} object of USER type.}

\item{charactes}{A numeric vector of characters to be set as inactive.}
}
\value{
A \code{phyDat} object.
}
\description{
Define inactive characters
}
