% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_support.R
\name{new_support}
\alias{new_support}
\title{New support method configuration}
\usage{
new_support(name, ...)
}
\arguments{
\item{name}{The name of a support method. Partial unambiguous matches are
also accepted. The valid options are:
\itemize{
\item \code{branch}: Calculates branch (i.e. Bremer) support statistics.
Attaches a \code{\link{BranchSupportOptions}} object.
\item \code{bootstrap}: Calculates bootstrap resampling support
statistics. Attaches a \code{\link{BootstrapOptions}} object.
\item \code{jackknife}: Calculates jackknife resmpling support
statistics. Attaches a \code{\link{JackknifeOptions}} object.
\item \code{symmetric}: Calculate symmetric resampling support
statistics. Attaches a \code{\link{SymmetricResamplingOptions}} object.
}}

\item{...}{Arguments to be passed on to the weighting method.}
}
\value{
A \code{\link{ResamplingOptions}} object.
}
\description{
Creates a new support method configuration.
}
\seealso{
The \code{new} method of \code{\link{BranchSupportOptions}},
\code{\link{BootstrapOptions}}, \code{\link{JackknifeOptions}} and
\code{\link{SymmetricResamplingOptions}}.
}
