% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_support.R
\name{new_support}
\alias{new_support}
\title{New support method configuration}
\usage{
new_support(name, ...)
}
\arguments{
\item{name}{The name of a support method. Partial unambiguous matches are
also accepted. The valid options are:
\itemize{
\item \code{branch}: Calculates branch (i.e. Bremer) support statistics.
Attaches a \code{\link{BranchSupportCommand}} object.
\item \code{bootstrap}: Calculates bootstrap resampling support
statistics. Attaches a \code{\link{ResamplingCommand}} object.
\item \code{jackknife}: Calculates jackknife resmpling support
statistics. Attaches a \code{\link{ResamplingCommand}} object.
\item \code{symmetric}: Calculate symmetric resampling support
statistics. Attaches a \code{\link{ResamplingCommand}} object.
}}

\item{...}{Arguments to be passed on to the weighting method.}
}
\value{
A \code{\link{ResamplingCommand}} object.
}
\description{
Creates a new support method configuration.
}
\seealso{
The \code{new} method of \code{\link{BranchSupportCommand}} and
\code{\link{ResamplingCommand}}.
}
