% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintsBase-methods.R
\name{newConstraint<-}
\alias{newConstraint<-}
\alias{newConstraint<-,NitroConstraintsBase,NitroConstraint-method}
\title{Add constraint}
\usage{
newConstraint(n) <- value

\S4method{newConstraint}{NitroConstraintsBase,NitroConstraint}(n) <- value
}
\arguments{
\item{value}{an object that inherits from \code{\link{NitroConstraint}}}
}
\description{
Function to add a new \code{\link{NitroConstraint}} to a
\code{\link{NitroConstraintBase}} object.
}
