% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{newTreeSearch}
\alias{newTreeSearch}
\title{Define tree search parameters}
\usage{
newTreeSearch(
  matrix,
  method,
  ordered_characters = numeric(),
  inactive_taxa = character(),
  inactive_characters = numeric(),
  outgroup = NULL,
  collapse = 3,
  weighting = c("equal", "implied"),
  k = 3,
  weights = numeric(),
  multi_k = FALSE,
  proportion = 0.5,
  max_ratio = 5
)
}
\arguments{
\item{matrix}{a \code{phyDat} object representing a character-taxon matrix}

\item{method}{an object that inherits from class
\code{"\linkS4class{NitroMethodsBase}"}}

\item{ordered_characters}{a numeric vector indicating the characters to be treated as ordered.}

\item{inactive_taxa}{a character vector indicating the taxa to be marked as inactive.}

\item{inactive_characters}{a numeric vector indicating the characters to be marked as inactive.}

\item{outgroup}{a single character vector indicating the taxon to be the outgroup.}

\item{collapse}{an integer indicating the rule for collapsing of zero length branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}

\item{weighting}{a character value indicating the type of character weighting to use during the analysis. The available options are \code{equal} (the default) or \code{implied}.}

\item{k}{an numeric value indicating the concavity constant to apply during an implied weights analysis. This parameter is only valid when \code{weighting} is set to \code{implied}. When \code{multi_k} is \code{TRUE}, \code{k} represents the concavity for a character with no missing entries and serves as the basis for calculating per-site concavity constants.}

\item{weights}{numeric}

\item{multi_k}{a logical value indicating whether each character will be given an independent concavity constant based on the value of \code{k} during an implied weights analysis. This parameter is only valid when \code{weighting} is set to \code{implied}.}

\item{proportion}{an numeric value indicating the proportion of homoplasy missing values are assumed to have under implied weighting. This parameter is only valid when \code{weighting} is set to \code{implied} and \code{multi_k} is \code{TRUE}. A proportion of 0 is equivalent to standard implied weighting, with \code{multi_k} set as \code{FALSE}.}

\item{max_ratio}{a numeric value indicating the maximum acceptable ratio between two k values when multiple concavity constants are used (i.e., when \code{multi_k} is \code{TRUE}). A maximum ratio of 1 is equivalent to standard implied weighting, with \code{multi_k} set as \code{FALSE}.}
}
\value{
an object of class \code{"\linkS4class{NitroTreeSearch}"}.
}
\description{
Define tree search parameters
}
