% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchBreak-class.R
\docType{class}
\name{NitroBranchBreak-class}
\alias{NitroBranchBreak-class}
\title{Define parameters for branch breaking}
\description{
\code{NitroBranchBreak} is an S4 class that defines the set of parameters
required to perform branch breaking.
}
\seealso{
The S4 classes \code{"\linkS4class{NitroImplicitEnum}"},
  \code{"\linkS4class{NitroBranchSwap}"}, \code{"\linkS4class{NitroRatchet}"} and
  \code{"\linkS4class{NitroDriven}"}.
}
\keyword{classes}
