% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchBreak-methods.R
\name{NitroBranchBreak}
\alias{NitroBranchBreak}
\title{Define parameters for branch breaking}
\usage{
NitroBranchBreak(
  swapper = c("tbr", "spr"),
  cluster_size = 15,
  safe_unclip = FALSE,
  fill_only = FALSE,
  save_multiple = TRUE,
  random_clip = FALSE
)
}
\arguments{
\item{swapper}{a character vector indicating the branch swapping algorithm
to use. Options are either '\code{tbr}' (the default) or '\code{spr}'.}

\item{cluster_size}{an integer value indicating the number of nodes
(clusters) to use during swapping. Applied only when \code{swapper} is
set to \code{tbr}. Larger clusters will result in faster swapping as
matrix size increases.}

\item{safe_unclip}{a logical value indicating whether to use a safer but
slower method for updating buffers when clipping to find a better tree.
Applied only when \code{swapper} is set to \code{tbr}.}

\item{fill_only}{a logical value indicating whether to stop swapping when
the tree buffer is full.}

\item{save_multiple}{a logical value indicating whether to save multiple
trees during swapping.}

\item{random_clip}{a logical value indicating whether to randomize the
tree clipping sequence.}
}
\description{
Define parameters for branch breaking
}
