% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchSwap-methods.R
\name{NitroBranchSwap}
\alias{NitroBranchSwap}
\title{Construct a branch swapping analysis}
\usage{
NitroBranchSwap(
  matrix,
  replications,
  hold_rep,
  keep_all = FALSE,
  ordered_characters = numeric(),
  inactive_taxa = character(),
  inactive_characters = numeric(),
  outgroup = NULL,
  collapse = 3,
  weighting = c("equal", "implied"),
  k = 3,
  multi_k = FALSE
)
}
\arguments{
\item{matrix}{an object of class \code{phyDat}.}

\item{replications}{an integer value indicating the number of replications.}

\item{hold_rep}{an integer value indicating the maximum number of trees to
retain during each replication.}

\item{keep_all}{a logical value indicating whether to retain all generated
trees from each replication regardless of length.}

\item{ordered_characters}{a numeric vector indicating the characters to be treated as ordered.}

\item{inactive_taxa}{a character vector indicating the taxa to be marked as inactive.}

\item{inactive_characters}{a numeric vector indicating the characters to be marked as inactive.}

\item{outgroup}{a single character vector indicating the taxon to be the outgroup.}

\item{collapse}{an integer indicating the rule for collapsing of zero length branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}

\item{weighting}{a character value indicating the type of character weighting to use during the analysis. The available options are \code{equal} (the default) or \code{implied}.}

\item{k}{an numeric value indicating the concavity constant to apply during an implied weights analysis. This parameter is only valid when \code{weighting} is set to \code{implied}.}

\item{multi_k}{a logical value indicating whether each character will be given an independent concavity constant based on the value of \code{k} during an implied weights analysis. This parameter is only valid when \code{weighting} is set to \code{implied}.}
}
\description{
Construct a branch swapping analysis
}
