% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroBranchSwap-methods.R
\name{NitroBranchSwap}
\alias{NitroBranchSwap}
\title{Define parameters for a branch swapping analysis}
\usage{
NitroBranchSwap(replications, hold_rep, keep_all = FALSE)
}
\arguments{
\item{replications}{an integer value indicating the number of replications.}

\item{hold_rep}{an integer value indicating the maximum number of trees to
retain during each replication.}

\item{keep_all}{a logical value indicating whether to retain all generated
trees from each replication regardless of length.}
}
\description{
Define parameters for a branch swapping analysis
}
