% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroConstraintSectorialSearch-methods.R
\name{NitroConstraintSectorialSearch}
\alias{NitroConstraintSectorialSearch}
\title{Define parameters for a constraint sectorial search analysis}
\usage{
NitroConstraintSectorialSearch(
  min_fork = 10,
  max_fork = 10,
  rounds = 3,
  buffer = TRUE,
  slack = 0
)
}
\arguments{
\item{min_fork}{an integer value indicating the minimum fork number to use
with constraint-based sectorial searches.}

\item{max_fork}{an integer value indicating the maximum fork number to use
with constraint-based sectorial searches.}

\item{rounds}{an integer value indicating the number of times to cycle over
groups in constraint-based selections.}

\item{buffer}{a logical value indicating whether to use an independent
memory buffer for analysis of sectors.}

\item{slack}{an integer value indicating the percentage to increase the
available memory during searches.}
}
\description{
Define parameters for a constraint sectorial search analysis
}
