% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroDriven-class.R
\docType{class}
\name{NitroDriven-class}
\alias{NitroDriven-class}
\title{Define a driven search}
\description{
\code{NitroBranchSwap} is an S4 class that defines the set of parameters
required to perform a driven search in \code{nitro}.
}
\seealso{
The inheriting S4 class \code{\link{NitroMethodsBase}} and
\code{\link{NitroImplicitEnum}}, \code{\link{NitroBranchSwap}},
\code{\link{NitroRatchet}} and \code{\link{Driven}}.
}
\keyword{classes}
