% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroDriven-methods.R
\name{NitroDriven}
\alias{NitroDriven}
\title{Definte parameters for a driven analysis}
\usage{
NitroDriven(
  replications = 4,
  hits = 1,
  consense_times = 0,
  keep_all = FALSE,
  multiply = TRUE,
  sectorial_search = NULL,
  tree_fuse = NULL,
  tree_hybridize = NULL,
  tree_drift = NULL,
  ratchet = NULL
)
}
\arguments{
\item{replications}{an integer value indicating the number of replications.}

\item{hits}{an integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{consense_times}{an integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{keep_all}{a logical value indicating whether to retain all generated
trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When
\code{hits} = 1, it is trees from each of the RAS + TBR +  SS or DFT or
RAT, in addition to the trees resulting from fusing those. When
\code{hits} > 1, then it means the trees resulting from fusing the
initial starting trees for each of starting points.}

\item{multiply}{a logical value indicating whether to find additional trees
by fusing suboptimal trees with optimal trees.}

\item{sectorial_search}{a list of objects of inheriting
\code{"\linkS4class{NitroSectorialSearch}"}.}

\item{tree_fuse}{an object of class \code{"\linkS4class{NitroTreeFuse}"}.}

\item{tree_hybridize}{an object of class \code{"\linkS4class{NitroTreeHybridize}"}.}

\item{tree_drift}{an object of class \code{"\linkS4class{NitroTreeDrift}"}.}

\item{ratchet}{an object of class \code{"\linkS4class{NitroRatchet}"}.}
}
\description{
Definte parameters for a driven analysis
}
