% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroDriven-methods.R
\name{NitroDriven}
\alias{NitroDriven}
\title{Construct a branch swapping analysis}
\usage{
NitroDriven(
  matrix,
  replications,
  hits = 1,
  rss = TRUE,
  css = FALSE,
  xss = FALSE,
  ratchet_cycles = 0,
  drifting_cycles = 30,
  fusing_rounds = 0,
  consense_times = 0,
  keep_all = FALSE,
  ordered_characters = numeric(),
  inactive_taxa = character(),
  inactive_characters = numeric(),
  outgroup = NULL,
  collapse = 3,
  weighting = c("equal", "implied"),
  k = 3,
  multi_k = FALSE
)
}
\arguments{
\item{matrix}{an object of class \code{phyDat}.}

\item{replications}{an integer value indicating the number of replications.}

\item{hits}{an integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{rss, css, xss}{a logical value indicating whether to use random,
constraint or exclusive sectorial searches.}

\item{ratchet_cycles}{an integer value indicating the number of cycles of
the parsimony ratchet to perform per replication.}

\item{drifting_cycles}{an integer value indicating the number of cycles of
tree drifting to perform per replication.}

\item{fusing_rounds}{an integer value indicating the number of rounds of
fusing per replication.}

\item{consense_times}{an integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{keep_all}{a logical value indicating whether to retain all generated
trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When
\code{hits} = 1, it is trees from each of the RAS + TBR +  SS or DFT or
RAT, in addition to the trees resulting from fusing those. When
\code{hits} > 1, then it means the trees resulting from fusing the
initial starting trees for each of starting points.}

\item{ordered_characters}{a numeric vector indicating the characters to be treated as ordered.}

\item{inactive_taxa}{a character vector indicating the taxa to be marked as inactive.}

\item{inactive_characters}{a numeric vector indicating the characters to be marked as inactive.}

\item{outgroup}{a single character vector indicating the taxon to be the outgroup.}

\item{collapse}{an integer indicating the rule for collapsing of zero length branches. The options are:
\itemize{
\item \code{1}: collapse an interior branch of the maximum possible length
  of the branch is zero
\item \code{2}: keep zero length branches if ancestor and descendent states
  differ
\item \code{3}: collapse an interior branch if the minimum possible length
  of the branch is zero (the default)
\item \code{4}: discard all trees that must contain a zero length branch
}}

\item{weighting}{a character value indicating the type of character weighting to use during the analysis. The available options are \code{equal} (the default) or \code{implied}.}

\item{k}{an numeric value indicating the concavity constant to apply during an implied weights analysis. This parameter is only valid when \code{weighting} is set to \code{implied}.}

\item{multi_k}{a logical value indicating whether each character will be given an independent concavity constant based on the value of \code{k} during an implied weights analysis. This parameter is only valid when \code{weighting} is set to \code{implied}.}
}
\description{
Construct a branch swapping analysis
}
