% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroDriven-class.R
\name{NitroDriven}
\alias{NitroDriven}
\title{Define a driven search}
\description{
\code{NitroBranchSwap} is an R6 class that defines the set of parameters
required to perform a driven search in \code{nitro}.
}
\section{Super class}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{NitroDriven}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{replications}}{An integer value indicating the number of replications.}

\item{\code{hits}}{An integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{\code{consense_times}}{An integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{\code{keep_all}}{A logical value indicating whether to retain all generated
trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When
\code{hits} = 1, it is trees from each of the RAS + TBR +  SS or DFT or
RAT, in addition to the trees resulting from fusing those. When
\code{hits} > 1, then it means the trees resulting from fusing the
initial starting trees for each of starting points.}

\item{\code{multiply}}{A logical value indicating whether to find additional trees
by fusing suboptimal trees with optimal trees.}

\item{\code{sectorial_search}}{A list of objects of inheriting
\code{"\link{NitroSectorialSearch}"}.}

\item{\code{tree_fuse}}{An object of class \code{"\link{NitroTreeFuse}"}.}

\item{\code{tree_hybridize}}{An object of class \code{"\link{NitroTreeHybridize}"}.}

\item{\code{tree_drift}}{An object of class \code{"\link{NitroTreeDrift}"}.}

\item{\code{ratchet}}{An object of class \code{"\link{NitroRatchet}"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroDriven$new()}}
\item \href{#method-print}{\code{NitroDriven$print()}}
\item \href{#method-tnt_cmd}{\code{NitroDriven$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroDriven$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroDriven$new(
  replications = 4,
  hits = 1,
  consense_times = 0,
  keep_all = FALSE,
  multiply = TRUE,
  sectorial_search = NULL,
  tree_fuse = NULL,
  tree_hybridize = NULL,
  tree_drift = NULL,
  ratchet = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replications}}{an integer value indicating the number of replications.}

\item{\code{hits}}{An integer value indicating the number of times the shortest
tree must be found on consecutive re-runs of the analysis before stopping.}

\item{\code{consense_times}}{An integer value indicating the number of times to
consense until the consensus is stablilised.}

\item{\code{keep_all}}{A logical value indicating whether to retain all generated
trees from each replication regardless of length. This has a different
meaning when \code{hits} = 1 and when \code{hits} > 1. When
\code{hits} = 1, it is trees from each of the RAS + TBR +  SS or DFT or
RAT, in addition to the trees resulting from fusing those. When
\code{hits} > 1, then it means the trees resulting from fusing the
initial starting trees for each of starting points.}

\item{\code{multiply}}{A logical value indicating whether to find additional trees
by fusing suboptimal trees with optimal trees.}

\item{\code{sectorial_search}}{A list of objects of inheriting
\code{"\link{NitroSectorialSearch}"}.}

\item{\code{tree_fuse}}{An object of class \code{"\link{NitroTreeFuse}"}.}

\item{\code{tree_hybridize}}{An object of class \code{"\link{NitroTreeHybridize}"}.}

\item{\code{tree_drift}}{An object of class \code{"\link{NitroTreeDrift}"}.}

\item{\code{ratchet}}{An object of class \code{"\link{NitroRatchet}"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroDriven$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroDriven$tnt_cmd(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroDriven$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
