% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroImplicitEnum-class.R
\docType{class}
\name{NitroImplicitEnum-class}
\alias{NitroImplicitEnum-class}
\title{Define an implicit enumeration analysis}
\description{
\code{NitroImplicitEnum} is an S4 class that defines an implicit enumeration
analysis in \code{nitro}.
}
\seealso{
The S4 classes \code{ImplicitEnum}, \code{Ratchet} and
\code{Driven}.
}
\keyword{classes}
