% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroImpliedWeights-methods.R
\name{NitroImpliedWeights}
\alias{NitroImpliedWeights}
\title{Define implied weights parameters}
\usage{
NitroImpliedWeights(k = 3, multi_k = FALSE, proportion = 0.5, max_ratio = 5)
}
\arguments{
\item{k}{n numeric value indicating the concavity constant to apply during
an implied weights analysis. When \code{multi_k} is \code{TRUE}, \code{k}
represents the concavity for a character with no missing entries and
serves as the basis for calculating per-site concavity constants.}

\item{multi_k}{a logical value indicating whether each character will be
given an independent concavity constant based on the value of \code{k}
during an implied weights analysis.}

\item{proportion}{an numeric value indicating the proportion of homoplasy
missing values are assumed to have under implied weighting. This parameter
is only valid when \code{multi_k} is \code{TRUE}. A proportion of 0 is
equivalent to standard implied weighting, with \code{multi_k} set as
\code{FALSE}.}

\item{max_ratio}{a numeric value indicating the maximum acceptable ratio
between two k values when multiple concavity constants are used (i.e.,
when \code{multi_k} is \code{TRUE}). A maximum ratio of 1 is equivalent
to standard implied weighting, with \code{multi_k} set as \code{FALSE}.}
}
\description{
Define implied weights parameters
}
