% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroRatchet-class.R
\docType{class}
\name{NitroRatchet-class}
\alias{NitroRatchet-class}
\title{Define a parsimony ratchet analysis}
\description{
\code{NitroRatchet} is an S4 class that defines the set of parameters
required to perform a branch swapping ('traditional', in TNTs terminology)
phylogenetic analysis in \code{nitro}.
}
\seealso{
The S4 classes \code{ImplicitEnum}, \code{Ratchet} and
\code{Driven}.
}
\keyword{classes}
