% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroSectorialSearch-class.R,
%   R/NitroConstraintSectorialSearch-class.R, R/NitroRatchet-class.R,
%   R/NitroTreeDrift-class.R, R/NitroTreeHybridize-class.R,
%   R/NitroTreeFuse-class.R, R/NitroRandomSectorialSearch-class.R
\docType{class}
\name{NitroRatchet-class}
\alias{NitroRatchet-class}
\title{Define a sectorial search analysis}
\description{
\code{NitroSectorialSearch} is an S4 class that defines the set of
parameters required to perform sectorial searching analyses in
\code{nitro}.

\code{NitroConstraintSectorialSearch} is an S4 class that defines the set of
parameters required to perform sectorial searching analyses in
\code{nitro}.

\code{NitroRatchet} is an S4 class that defines the set of parameters
required to perform a parsimony ratchet phylogenetic analysis in
\code{nitro}.

\code{NitroTreeDrift} is an S4 class that defines the set of parameters
required for performing tree drift operations in \code{nitro}.

\code{NitroTreeHybridize} is an S4 class that defines the set of parameters
required for performing tree hybridizing operations in \code{nitro}.

\code{NitroTreeFuse} is an S4 class that defines the set of parameters
required for performing tree fusing operations in \code{nitro}.

\code{NitroRandomSectorialSearch} is an S4 class that defines the set of
parameters required to perform sectorial searching analyses in
\code{nitro}.
}
\seealso{
The S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroRatchet}} and \code{\link{NitroDriven}}.

The S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroRatchet}} and \code{\link{NitroDriven}}.

The S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroRatchet}} and \code{\link{NitroDriven}}.

The S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroRatchet}} and \code{\link{NitroDriven}}.

The S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroRatchet}} and \code{\link{NitroDriven}}.

The S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroRatchet}} and \code{\link{NitroDriven}}.

The S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroRatchet}} and \code{\link{NitroDriven}}.
}
\keyword{classes}
