% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroRatchet-class.R
\name{NitroRatchet}
\alias{NitroRatchet}
\title{Define a parsimony ratchet analysis}
\description{
\code{NitroRatchet} is an R6 class that defines the set of parameters
required to perform a parsimony ratchet phylogenetic analysis in
\code{nitro}.
}
\section{Super class}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{NitroRatchet}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iterations}}{An integer value indicating the number of iterations.}

\item{\code{replacements}}{An integer value indicating the number of
replacements (i.e., accepted tree rearrangements) to perform in each
perturbation phase.}

\item{\code{prob_up}}{An integer value indicating the probability of
upweighting a character.}

\item{\code{prob_down}}{An integer value indicating the probability of
downweighting a character.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroRatchet$new()}}
\item \href{#method-print}{\code{NitroRatchet$print()}}
\item \href{#method-tnt_cmd}{\code{NitroRatchet$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroRatchet$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroRatchet$new(
  iterations = 50,
  replacements = 40,
  prob_up = 4,
  prob_down = 4
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{An integer value indicating the number of iterations.}

\item{\code{replacements}}{An integer value indicating the number of
replacements (i.e., accepted tree rearrangements) to perform in each
perturbation phase.}

\item{\code{prob_up}}{An integer value indicating the probability of
upweighting a character.}

\item{\code{prob_down}}{An integer value indicating the probability of
downweighting a character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroRatchet$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroRatchet$tnt_cmd(set_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_only}}{A logical indicating whether to instruct the command to
execute immediately (\code{TRUE}) or set the variables for future
execution (\code{FALSE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroRatchet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
