% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeDrift-class.R
\name{NitroTreeDrift}
\alias{NitroTreeDrift}
\title{Define tree drift properties}
\description{
\code{NitroTreeDrift} is an R6 class that defines the set of parameters
required for performing tree drift operations in \code{nitro}.
}
\section{Super class}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{NitroTreeDrift}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iterations}}{An integer value indicating the number of tree-drifting
cycles to perform.}

\item{\code{substitutions}}{An integer value indicating the number of replacements
(i.e., accepted tree rearrangements) to perform in the perturbation phase.}

\item{\code{max_abs_fit_diff}}{A numeric value indicating the maximum absolute fit
difference.}

\item{\code{max_rel_fit_diff}}{A numeric value indicating the maximum relative fit
difference.}

\item{\code{reject_factor}}{A numeric value indicating the rejection factor for
suboptimal trees.}

\item{\code{autoconstrain_cycles}}{An integer value indicating the number of
autoconstrained cycles to perform.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroTreeDrift$new()}}
\item \href{#method-print}{\code{NitroTreeDrift$print()}}
\item \href{#method-tnt_cmd}{\code{NitroTreeDrift$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroTreeDrift$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeDrift$new(
  iterations = 30,
  substitutions = 60,
  max_abs_fit_diff = 1,
  max_rel_fit_diff = 0.2,
  reject_factor = 3,
  autoconstrain_cycles = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{An integer value indicating the number of tree-drifting
cycles to perform.}

\item{\code{substitutions}}{An integer value indicating the number of replacements
(i.e., accepted tree rearrangements) to perform in the perturbation phase.}

\item{\code{max_abs_fit_diff}}{A numeric value indicating the maximum absolute fit
difference.}

\item{\code{max_rel_fit_diff}}{A numeric value indicating the maximum relative fit
difference.}

\item{\code{reject_factor}}{A numeric value indicating the rejection factor for
suboptimal trees.}

\item{\code{autoconstrain_cycles}}{An integer value indicating the number of
autoconstrained cycles to perform.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeDrift$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeDrift$tnt_cmd(set_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_only}}{A logical indicating whether to instruct the command to
execute immediately (\code{TRUE}) or set the variables for future
execution \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeDrift$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
