% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeFuse-class.R
\name{NitroTreeFuse}
\alias{NitroTreeFuse}
\title{Define tree fusing properties}
\description{
\code{NitroTreeFuse} is an R6 class that defines the set of parameters
required for performing tree fusing operations in \code{nitro}.
}
\section{Super class}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{NitroTreeFuse}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of tree-fusing rounds
to perform.}

\item{\code{exchange_equal}}{A logical value indicating whether to accept exchanges
of equal score.}

\item{\code{start_best}}{A logical value indicating whether to use the best tree to
start tree-fusing.}

\item{\code{keep_all}}{A logical value indicating whether to keep all trees found
instead of only the best trees.}

\item{\code{accept_all}}{A logical value indicating whether to accept all exchanges
rather than only those that improve the best score.}

\item{\code{swap}}{A logical value indicating whether to perform tree-bisection
reconnection swapping after exchanging clades.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroTreeFuse$new()}}
\item \href{#method-print}{\code{NitroTreeFuse$print()}}
\item \href{#method-tnt_cmd}{\code{NitroTreeFuse$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroTreeFuse$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeFuse$new(
  rounds = 5,
  exchange_equal = FALSE,
  start_best = TRUE,
  keep_all = TRUE,
  accept_all = TRUE,
  swap = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of tree-fusing rounds
to perform.}

\item{\code{exchange_equal}}{A logical value indicating whether to accept exchanges
of equal score.}

\item{\code{start_best}}{A logical value indicating whether to use the best tree to
start tree-fusing.}

\item{\code{keep_all}}{A logical value indicating whether to keep all trees found
instead of only the best trees.}

\item{\code{accept_all}}{A logical value indicating whether to accept all exchanges
rather than only those that improve the best score.}

\item{\code{swap}}{A logical value indicating whether to perform tree-bisection
reconnection swapping after exchanging clades.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeFuse$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeFuse$tnt_cmd(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeFuse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
