% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeHybridize-class.R
\name{NitroTreeHybridize}
\alias{NitroTreeHybridize}
\title{Define tree hybridization properties}
\description{
\code{NitroTreeHybridize} is an R6 class that defines the set of parameters
required for performing tree hybridizing operations in \code{nitro}.
}
\section{Super class}{
\code{\link[nitro:NitroMethodsBase]{nitro::NitroMethodsBase}} -> \code{NitroTreeHybridize}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of rounds of
tree-hybridizing to perform.}

\item{\code{hybridizations}}{An integer value indicating the number of
hybridizations to perform in each round.}

\item{\code{best_trees}}{An integer value indicating the number of best trees from
the previous round of hybridizing to use in the next round.}

\item{\code{replace}}{A logical value indicating whether to replace the source tree
with a better tree produced by hybridizing.}

\item{\code{sample_factor}}{An integer value indicating the number of times to
increase the size of initial tree set by. The corresponding number of trees
to retain will be proportional to the inverse of this value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NitroTreeHybridize$new()}}
\item \href{#method-print}{\code{NitroTreeHybridize$print()}}
\item \href{#method-tnt_cmd}{\code{NitroTreeHybridize$tnt_cmd()}}
\item \href{#method-clone}{\code{NitroTreeHybridize$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeHybridize$new(
  rounds = 1,
  hybridizations = 1000,
  best_trees = 50,
  replace = TRUE,
  sample_factor = 15
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rounds}}{An integer value indicating the number of rounds of
tree-hybridizing to perform.}

\item{\code{hybridizations}}{An integer value indicating the number of
hybridizations to perform in each round.}

\item{\code{best_trees}}{An integer value indicating the number of best trees from
the previous round of hybridizing to use in the next round.}

\item{\code{replace}}{A logical value indicating whether to replace the source tree
with a better tree produced by hybridizing.}

\item{\code{sample_factor}}{An integer value indicating the number of times to
increase the size of initial tree set by. The corresponding number of trees
to retain will be proportional to the inverse of this value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeHybridize$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tnt_cmd"></a>}}
\if{latex}{\out{\hypertarget{method-tnt_cmd}{}}}
\subsection{Method \code{tnt_cmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeHybridize$tnt_cmd(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NitroTreeHybridize$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
