% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-class.R
\docType{class}
\name{NitroTreeSearch-class}
\alias{NitroTreeSearch-class}
\title{Define tree searches}
\description{
\code{NitroBase} is an S4 class that serves as the basis for classes that
define parameters for equal and implied weights analyses.
}
\seealso{
The inheriting S4 classes \code{NitroEqualWeights} and
\code{NitroImpliedWeights}.
}
\keyword{classes}
