% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroMethodsBase-class.R,
%   R/NitroWeightsBase-class.R, R/NitroConstraintsBase-class.R,
%   R/NitroTreeSearch-class.R
\docType{class}
\name{NitroTreeSearch-class}
\alias{NitroTreeSearch-class}
\title{Define tree searches}
\description{
\code{NitroMethodsBase} is an S4 class that serves as the basis for classes that
define parameters for tree search methods.

\code{NitroWeighBase} is an S4 class that serves as the basis for classes
that define parameters for equal and implied weights analyses.

\code{NitroConstraintsBase} is an S4 class that serves as the basis for classes
that define parameters for constraints on monophyly.

\code{NitroTreeSearch} is an S4 class that stores information regarding the
tree search method, character weighting scheme and constraints on monophyly.
}
\seealso{
The inheriting S4 classes \code{\link{NitroImplicitEnum}},
\code{\link{NitroBranchSwap}}, \code{\link{NitroRatchet}} and
\code{\link{NitroDriven}}.

The inheriting S4 classes \code{\link{NitroEqualWeights}} and
\code{\link{NitroImpliedWeights}}.

The inheriting S4 classes \code{\link{NitroUnconstrainedSearch}}
and \code{\link{NitroConstrainedSearch}}.

The S4 classes \code{\link{NitroConstraintsBase}},
\code{\link{NitroMethodsBase}} and \code{\link{NitroWeightsBase}}.
}
\keyword{classes}
