% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractCharacterMatrix-class.R
\name{print.MultiCharacterMatrix}
\alias{print.MultiCharacterMatrix}
\title{Print a MultiCharacterMatrix Object}
\usage{
\method{print}{MultiCharacterMatrix}(x, ...)
}
\arguments{
\item{x}{A \code{MultiCharacterMatrix} object.}

\item{...}{Ignored.}
}
\description{
Print a MultiCharacterMatrix Object
}
\keyword{internal}
