% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raven_2017}
\alias{raven_2017}
\title{Raven and Maidment (2017) matrix}
\format{
A list containing the following objects:
  \itemize{
    \item \code{continuous}: A data frame containing 24 continuous character
      traits; and
    \item \code{discrete}: A phyDat object containing 91 characters.
  }
  Both objects score character data for the same 23 taxa.
}
\source{
Downloaded from the Dryad Digital repository of Raven and
  Maidment (2017) <https://doi.org/10.5061/dryad.ds543>. The TNT file was
  converted into both a CSV file for the continuous data, and a Nexus file in
  Mesquite 3.61 for the discrete character data. The CSV file was read with
  `read.table` and the Nexus file was read with ReadAsPhydat from TreeTools.
}
\usage{
raven_2017
}
\description{
A phylogenetic matrix of stegosaurian dinosaurs.
}
\keyword{datasets}
