% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTnt.R, R/runTnt2.R
\name{runTnt}
\alias{runTnt}
\title{Executes TNT commands}
\usage{
runTnt(tnt.path, matrix, tnt.params, timeout = 10000)

runTnt(tnt.path, matrix, tnt.params, timeout = 10000)
}
\arguments{
\item{tnt.path}{The location of the TNT command line binary.}

\item{matrix}{A \code{phyDat} object of the matrix.}

\item{tnt.params}{A list of parameters for executing commands in TNT.}

\item{timeout}{Time after which to terminate a non-responsive TNT process
(milliseconds).}

\item{analysis}{The parameters for an analysis as specified from the output
of the \code{branchswap}, \code{ratchet}, or \code{driven} commands.}

\item{tnt.path}{The location of the TNT command line binary.}

\item{timeout}{Time after which to terminate a non-responsive TNT process
(milliseconds).}
}
\value{
A character vector of the output from TNT.

A character vector of the output from TNT.
}
\description{
Executes commands on a supplied phylogenetic matrix in the TNT
  command line binary.

Executes commands on a supplied phylogenetic matrix in the TNT
  command line binary.
}
