% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectorialSearchBaseOptions-class.R
\name{SectorialSearchBaseOptions}
\alias{SectorialSearchBaseOptions}
\title{Set options common to sectorial searches}
\description{
\code{SectorialSearchBaseOptions} is an R6 class that defines the set of
  options common to all types of sectorial searches in \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{buffer}}{A logical value indicating whether to use an independent}

\item{\code{slack}}{An integer value indicating the percentage to increase the
available memory during searches.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SectorialSearchBaseOptions-new}{\code{SectorialSearchBaseOptions$new()}}
\item \href{#method-SectorialSearchBaseOptions-clone}{\code{SectorialSearchBaseOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectorialSearchBaseOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-SectorialSearchBaseOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectorialSearchBaseOptions$new(buffer = TRUE, slack = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buffer}}{A logical value indicating whether to use an independent}

\item{\code{slack}}{An integer value indicating the percentage to increase the
available memory during searches.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectorialSearchBaseOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SectorialSearchBaseOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectorialSearchBaseOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
