% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R
\name{inactive_taxa}
\alias{inactive_taxa}
\alias{inactive_taxa,NitroTreeSearch-method}
\alias{inactive_taxa<-}
\alias{inactive_taxa<-,NitroTreeSearch,character-method}
\alias{inactive_taxa<-,NitroTreeSearch,NULL-method}
\title{List inactive taxa}
\usage{
inactive_taxa(n)

\S4method{inactive_taxa}{NitroTreeSearch}(n)

inactive_taxa(n) <- value

\S4method{inactive_taxa}{NitroTreeSearch,character}(n) <- value

\S4method{inactive_taxa}{NitroTreeSearch,`NULL`}(n) <- value
}
\arguments{
\item{n}{a \code{NitroTreeSearch} object}

\item{value}{a character vector indicating the taxa to be marked as inactive.}
}
\value{
a character vector indicating the taxa defined as inactive.
}
\description{
Function that returns the names of taxa defined as inactive.
}
