% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt.R
\name{tnt}
\alias{tnt}
\title{Executes TNT commands}
\usage{
tnt(obj, tnt_path, hold, max_ram = 16)
}
\arguments{
\item{obj}{an object that inherits from \code{NitroBase} (i.e.,
\code{IterativeEnum}, \code{BranchSwap}, \code{Ratchet} or \code{Driven}.}

\item{tnt_path}{the location of the TNT command line binary.}

\item{hold}{an integer indicating the number of trees to hold in TNTs tree
buffer.}

\item{max_ram}{a numeric indicating the number of (binary) megabytes to
allocate for use by TNT.}
}
\value{
An object containing the TNT parameters and the \code{phyDat} matrix
  used in the analysis, and a \code{multiPhylo} containing the trees foundas()
  during the search.
}
\description{
Executes commands on a supplied phylogenetic matrix in the TNT
  command line binary.
}
