% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt.R
\name{tnt}
\alias{tnt}
\title{Execute a TNT search}
\usage{
tnt(obj, tnt_path, read_trees = FALSE, character_fits = FALSE)
}
\arguments{
\item{obj}{an object that inherits \code{"\linkS4class{NitroTreeSearch}"}.}

\item{tnt_path}{the location of the TNT command line binary.}

\item{read_trees}{a logical value indicating whether to read in trees from
the \code{start_trees} slot of \code{obj} prior to analysis.}

\item{character_fits}{a logical value indicating whether to return scores
for individual characters and minimum and maximum possible character
lengths required for calculation of character fit indices.}
}
\value{
a \code{"\linkS4class{NitroTrees}"} object containing the original
  \code{"\linkS4class{NitroTreeSearch}"} object and a \code{multiPhylo} of
  the trees found during the search.
}
\description{
Executes commands on a supplied phylogenetic matrix in the TNT
  command line binary.
}
\details{
This function uses \code{\link{writeTNTNexus}} to generate a nexus file
containing the commands required to conduct a tree search using the matrix
and parameters specified within a \code{"\linkS4class{NitroTreeSearch}"}
object.
}
