% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt.R
\name{tnt}
\alias{tnt}
\title{Executes TNT commands}
\usage{
tnt(obj, tnt_path, hold, max_ram = 16, read_trees = FALSE)
}
\arguments{
\item{obj}{an object that inherits \code{NitroTreeSearch}.}

\item{tnt_path}{the location of the TNT command line binary.}

\item{hold}{an integer indicating the number of trees to hold in TNTs tree
buffer.}

\item{max_ram}{a numeric indicating the number of (binary) megabytes to
allocate for use by TNT.}

\item{read_trees}{a logical value indicating whether to read in trees from
a \code{\link{NitroResults}} object prior to analysis.}
}
\value{
the original \code{NitroTreeSearch} object containing a
  \code{NitroResults} object containing the trees found during the search.
}
\description{
Executes commands on a supplied phylogenetic matrix in the TNT
  command line binary.
}
