% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTnt.R
\name{tnt}
\alias{tnt}
\title{Executes TNT commands}
\usage{
tnt(tnt.path, analysis, progress)
}
\arguments{
\item{tnt.path}{The location of the TNT command line binary.}

\item{analysis}{The parameters for an analysis as specified from the output
of the \code{branchswap}, \code{ratchet}, or \code{driven} commands.}

\item{progress}{A list providing the progress bar object and a callback
function for updating the progress bar.}
}
\value{
A character vector of the output from TNT.
}
\description{
Executes commands on a supplied phylogenetic matrix in the TNT
  command line binary.
}
