% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NitroTreeSearch-methods.R,
%   R/NitroBranchBreak-methods.R, R/NitroBranchSwap-methods.R,
%   R/NitroConstraint-methods.R, R/NitroConstraintSectorialSearch-methods.R,
%   R/NitroConstraintsBase-methods.R, R/NitroDriven-methods.R,
%   R/NitroImplicitEnum-methods.R, R/NitroImpliedWeights-methods.R,
%   R/NitroJacknife-methods.R, R/NitroRandomSectorialSearch-methods.R,
%   R/NitroRatchet-methods.R, R/NitroTreeDrift-methods.R,
%   R/NitroTreeFuse-methods.R, R/NitroTreeHybridize-methods.R,
%   R/NitroTrees-methods.R
\name{tnt_cmd}
\alias{tnt_cmd}
\alias{tnt_cmd,NitroBranchBreak-method}
\alias{tnt_cmd,NitroBranchSwap-method}
\alias{tnt_cmd,NitroConstraint-method}
\alias{tnt_cmd,NitroConstraintSectorialSearch-method}
\alias{tnt_cmd,NitroConstraintsBase-method}
\alias{tnt_cmd,NitroDriven-method}
\alias{tnt_cmd,NitroImplicitEnum-method}
\alias{tnt_cmd,NitroImpliedWeights-method}
\alias{tnt_cmd,NitroJacknife-method}
\alias{tnt_cmd,NitroRandomSectorialSearch-method}
\alias{tnt_cmd,NitroRatchet-method}
\alias{tnt_cmd,NitroTreeDrift-method}
\alias{tnt_cmd,NitroTreeFuse-method}
\alias{tnt_cmd,NitroTreeHybridize-method}
\alias{tnt_cmd,NitroTrees-method}
\title{Return TNT command}
\usage{
tnt_cmd(n)

\S4method{tnt_cmd}{NitroBranchBreak}(n)

\S4method{tnt_cmd}{NitroBranchSwap}(n)

\S4method{tnt_cmd}{NitroConstraint}(n)

\S4method{tnt_cmd}{NitroConstraintSectorialSearch}(n)

\S4method{tnt_cmd}{NitroConstraintsBase}(n)

\S4method{tnt_cmd}{NitroDriven}(n)

\S4method{tnt_cmd}{NitroImplicitEnum}(n)

\S4method{tnt_cmd}{NitroImpliedWeights}(n)

\S4method{tnt_cmd}{NitroJacknife}(n)

\S4method{tnt_cmd}{NitroRandomSectorialSearch}(n)

\S4method{tnt_cmd}{NitroRatchet}(n)

\S4method{tnt_cmd}{NitroTreeDrift}(n)

\S4method{tnt_cmd}{NitroTreeFuse}(n)

\S4method{tnt_cmd}{NitroTreeHybridize}(n)

\S4method{tnt_cmd}{NitroTrees}(n)
}
\arguments{
\item{n}{an object inheriting from \code{"\linkS4class{NitroTreeSearch}"}.}
}
\value{
a character vector of the TNT command.
}
\description{
A function that returns the command to perform the phylogenetic analysis for
a branch swapping analysis.
}
