% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntTreeParse.R
\name{tntTreeParse}
\alias{tntTreeParse}
\title{Extracts trees from TNT output.}
\usage{
tntTreeParse(tnt.output, tip.labels)
}
\arguments{
\item{tnt.output}{TNT output.}

\item{tip.labels}{A list or vector containing the names of the tips.}
}
\value{
A \code{multiphylo} object.
}
\description{
Extracts trees from TNT output.
}
