% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeAnalysis-class.R
\name{TreeAnalysis}
\alias{TreeAnalysis}
\title{Define tree searches}
\description{
\code{TreeAnalysis} is an R6 class that stores the configuration for a
 tree analysis, including the character-taxon matrix, taxon activity,
 character weighting and constraints on monophyly.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{discrete_matrix}}{A \code{"\link{DiscreteMatrix}"} object.}

\item{\code{continuous_matrix}}{A \code{"\link{ContinuousMatrix}"} object.}

\item{\code{method}}{An object that contains configuration options for a tree
analysis method.}

\item{\code{inactive_taxa}}{A character vector indicating the taxa to be
inactivated.}

\item{\code{outgroup}}{A single character vector indicating the taxon to be
the outgroup.}

\item{\code{zlb_rule}}{A character vector indicating the rule for handling zero
length branches. The options are:
\itemize{
\item \code{maximum}: collapse an interior branch of the maximum possible
  length of the branch is zero;
\item \code{identical_states}: only collapse zero length branches if ancestor and descendant
  states are the same;
\item \code{minimum}: collapse an interior branch if the minimum possible
  length of the branch is zero (default);
\item \code{discard_tree}: discard all trees that must contain a zero length
  branch;
\item \code{spr}: collapse an interior branch using subtree pruning and reconnection (SPR) operations; and
\item \code{tbr}: collapse an interior branch using tree bisection and reconnection (TBR) operations.
}}

\item{\code{constraints}}{One or more \code{"\link{MonophylyConstraintOptions}"} objects.}

\item{\code{weights}}{An object containing configuration options for character
weighting.}

\item{\code{start_trees}}{A \code{phylo} or \code{multiPhylo} of trees to load
prior to starting the tree analysis.}

\item{\code{hold}}{An integer indicating the number of trees to hold in TNT's
tree buffer.}

\item{\code{max_ram}}{A numeric indicating the number of (binary) megabytes to
allocate for use by TNT.}

\item{\code{timeout}}{A positive integer indicating the number of seconds to
allow a search to run for before terminating.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreeAnalysis-new}{\code{TreeAnalysis$new()}}
\item \href{#method-TreeAnalysis-print}{\code{TreeAnalysis$print()}}
\item \href{#method-TreeAnalysis-queue}{\code{TreeAnalysis$queue()}}
\item \href{#method-TreeAnalysis-run}{\code{TreeAnalysis$run()}}
\item \href{#method-TreeAnalysis-clone}{\code{TreeAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysis-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysis$new(
  method,
  discrete_matrix = NULL,
  continuous_matrix = NULL,
  inactive_taxa = NULL,
  outgroup = NULL,
  zlb_rule = "minimum",
  constraints = NULL,
  weights = NULL,
  start_trees = NULL,
  hold = 100,
  max_ram = 16,
  timeout = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{An object that contains configuration options for the tree
analysis method.}

\item{\code{discrete_matrix}}{A \code{"\link{DiscreteMatrix}"} object.}

\item{\code{continuous_matrix}}{A \code{"\link{ContinuousMatrix}"} object.}

\item{\code{inactive_taxa}}{A character vector indicating the taxa to be
inactivated.}

\item{\code{outgroup}}{A single character vector indicating the taxon to be the
outgroup.}

\item{\code{zlb_rule}}{An integer indicating the rule for collapsing of zero
length branches. The options are:
\itemize{
\item \code{maximum}: collapse an interior branch of the maximum
 possible length of the branch is zero;
\item \code{identical_states}: only collapse zero length branches if
 ancestor and descendant states are the same;
\item \code{minimum}: collapse an interior branch if the minimum
 possible length of the branch is zero (the default); and
\item \code{discard_tree}: discard all trees that must contain a zero
 length branch.
}}

\item{\code{constraints}}{One or more \code{"\link{MonophylyConstraintOptions}"} objects.}

\item{\code{weights}}{An object containing configuration options for character
weighting.}

\item{\code{start_trees}}{A \code{phylo} or \code{multiPhylo} of trees to load
prior to starting the tree analysis.}

\item{\code{hold}}{An integer indicating the number of trees to hold in TNTs tree
buffer.}

\item{\code{max_ram}}{A numeric indicating the number of (binary) megabytes to
allocate for use by TNT.}

\item{\code{timeout}}{A positive integer indicating the number of seconds to
allow a search to run for before terminating.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysis-print"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysis-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysis$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysis-queue"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysis-queue}{}}}
\subsection{Method \code{queue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysis$queue(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysis-run"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysis-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysis$run(.envir = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.envir}}{The environment that TNT has been attached to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
