% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeAnalysisResults-class.R
\name{TreeAnalysisResults}
\alias{TreeAnalysisResults}
\title{Results of a tree analyses}
\description{
\code{TreeAnalysisResults} is an R6 class that contains the results of a tree
  analysis in \code{nitro}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{trees}}{The trees generated from the analysis of \code{queue}.}

\item{\code{queue}}{The \code{CommandQueue} object and generated the results.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreeAnalysisResults-new}{\code{TreeAnalysisResults$new()}}
\item \href{#method-TreeAnalysisResults-print}{\code{TreeAnalysisResults$print()}}
\item \href{#method-TreeAnalysisResults-clone}{\code{TreeAnalysisResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysisResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysisResults-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysisResults$new(
  phy,
  queue,
  lengths = NULL,
  adjusted_homoplasy_scores = NULL,
  min_length = NULL,
  max_length = NULL,
  tags = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phy}}{A \code{multiPhylo} object.}

\item{\code{queue}}{A \code{"\link{CommandQueue}"} object.}

\item{\code{lengths}}{A numeric vector of lengths for \code{trees}.}

\item{\code{adjusted_homoplasy_scores}}{A numeric vector of adjusted homoplasy scores for \code{trees}.}

\item{\code{min_length}}{A number indicating the minimum possible length for
\code{trees} given the active characters and taxa.}

\item{\code{max_length}}{A number indicating the maximum possible length for
\code{trees} given the active characters and taxa.}

\item{\code{tags}}{A data frame containing annotation variables for \code{trees}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysisResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysisResults-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysisResults$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeAnalysisResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreeAnalysisResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreeAnalysisResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
