% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBufferCommand-class.R
\name{TreeBufferCommand}
\alias{TreeBufferCommand}
\title{Tree Buffer Command}
\description{
An \link[R6:R6Class]{R6} class that allocates memory to the TNT tree buffer in
terms of the maximum number of trees to hold simultaneously.

This command is created automatically as part of the standard
\link{CommandQueue} setup. Users do not typically need to instantiate it
directly.
}
\details{
\subsection{Command output}{

\verb{$render()} produces a string of the form \verb{hold \{size\};}
}

\subsection{Queue integration}{

Calling \verb{$enqueue()} adds this command to a \link{CommandQueue} at priority
\code{400}.
}
}
\keyword{internal}
\section{Super class}{
\code{\link[nitro:BasicCommand]{BasicCommand}} -> \code{TreeBufferCommand}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{size}}{[\code{integer(1)}]\cr
The maximum number of trees to allocate in the TNT tree buffer. Must
be a positive integer.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TreeBufferCommand-enqueue}{\code{TreeBufferCommand$enqueue()}}
    \item \href{#method-TreeBufferCommand-initialize}{\code{TreeBufferCommand$new()}}
    \item \href{#method-TreeBufferCommand-render}{\code{TreeBufferCommand$render()}}
    \item \href{#method-TreeBufferCommand-clone}{\code{TreeBufferCommand$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="format"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-format'><code>BasicCommand$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="get_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-get_dependency'><code>BasicCommand$get_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="new_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-new_dependency'><code>BasicCommand$new_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="print"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-print'><code>BasicCommand$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="set_dependency"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-set_dependency'><code>BasicCommand$set_dependency()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="nitro" data-topic="BasicCommand" data-id="transform"><a href='../../nitro/html/BasicCommand.html#method-BasicCommand-transform'><code>BasicCommand$transform()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeBufferCommand-enqueue"></a>}}
\if{latex}{\out{\hypertarget{method-TreeBufferCommand-enqueue}{}}}
\subsection{\code{TreeBufferCommand$enqueue()}}{
  Add this command to a \link{CommandQueue}.

Adds this command at priority \code{400}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeBufferCommand$enqueue(.queue = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{.queue}}{A \link{CommandQueue} object, or \code{NULL} to create a new
one.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \link{CommandQueue} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeBufferCommand-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TreeBufferCommand-initialize}{}}}
\subsection{\code{TreeBufferCommand$new()}}{
  Create a new \code{TreeBufferCommand} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeBufferCommand$new(size)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{size}}{[\code{integer(1)}]\cr
The number of trees to allocate in the tree buffer. See the
\verb{$size} field.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{TreeBufferCommand} object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeBufferCommand-render"></a>}}
\if{latex}{\out{\hypertarget{method-TreeBufferCommand-render}{}}}
\subsection{\code{TreeBufferCommand$render()}}{
  Render the TNT \code{hold} command string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeBufferCommand$render(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Not used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A single-element character vector containing the TNT
command.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreeBufferCommand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreeBufferCommand-clone}{}}}
\subsection{\code{TreeBufferCommand$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TreeBufferCommand$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
